% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{aic.climex.fit.gpd}
\alias{aic.climex.fit.gpd}
\title{Akaike information criterion}
\usage{
\method{aic}{climex.fit.gpd}(x)
}
\arguments{
\item{x}{An object returned by the \code{\link{fit.gpd}} function.}
}
\value{
Numeric value
}
\description{
Calculates the Akaike information criterion of a
  \emph{climex.fit.gpd} class object.
}
\details{
The \emph{climex.fit.gpd} object is of identical
  structure as the output of the \code{\link[stats]{optim}}
  function.
}
\seealso{
\code{\link{bic}}, \code{\link{fit.gpd}}

Other ts: \code{\link{aic.climex.fit.gev}},
  \code{\link{aic.list}}, \code{\link{aic}},
  \code{\link{anomalies.list}},
  \code{\link{anomalies.xts}}, \code{\link{anomalies}},
  \code{\link{bic.climex.fit.gev}},
  \code{\link{bic.climex.fit.gpd}}, \code{\link{bic.list}},
  \code{\link{bic}}, \code{\link{check.completeness}},
  \code{\link{remove.incomplete.years.list}},
  \code{\link{remove.incomplete.years.xts}},
  \code{\link{remove.incomplete.years}},
  \code{\link{remove.seasonality.list}},
  \code{\link{remove.seasonality.xts}},
  \code{\link{remove.seasonality}}
}
\author{
Philipp Mueller
}
\concept{ts}
