% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{aic.list}
\alias{aic.list}
\title{Akaike information criterion}
\usage{
\method{aic}{list}(x)
}
\arguments{
\item{x}{Either a list of objects returned by the
\code{\link{fit.gev}} or \code{\link{fit.gpd}} function.}
}
\value{
Numerical vector
}
\description{
Calculates the Akaike information criterion of a
  list of \emph{climex.fit.gev} or \emph{climex.fit.gpd} class
  objects.
}
\details{
The climex.fit.gev object is of identical structure as
  the output of the \code{\link{optim}} function.
}
\seealso{
\code{\link{bic}}, \code{\link{fit.gev}},
  \code{\link{fit.gpd}}

Other ts: \code{\link{aic.climex.fit.gev}},
  \code{\link{aic}}, \code{\link{anomalies.list}},
  \code{\link{anomalies.xts}}, \code{\link{anomalies}},
  \code{\link{bic.climex.fit.gev}}, \code{\link{bic.list}},
  \code{\link{bic}}, \code{\link{check.completeness}},
  \code{\link{remove.incomplete.years.list}},
  \code{\link{remove.incomplete.years.xts}},
  \code{\link{remove.incomplete.years}},
  \code{\link{remove.seasonality.list}},
  \code{\link{remove.seasonality.xts}},
  \code{\link{remove.seasonality}}
}
\author{
Philipp Mueller
}
