% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{aic}
\alias{aic}
\title{Akaike information criterion}
\usage{
aic(x)
}
\arguments{
\item{x}{Optim() like fit of a time series.}
}
\value{
Numeric value
}
\description{
Calculates the Akaike information criterion of a climex.fit.gev or climex.fit.gpd object.
}
\details{
The climex.fit.gev object is of identical structure as the output of the optim() function.
}
\seealso{
\code{\link{bic}}

Other ts: \code{\link{anomalies}}, \code{\link{bic}},
  \code{\link{check.completeness}},
  \code{\link{remove.incomplete.years}},
  \code{\link{remove.seasonality}}
}
\author{
Philipp Mueller
}
