% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules2.R
\name{animation.wrapper}
\alias{animation.wrapper}
\title{Help function filling the JavaScript template which produces the likelihood animation.}
\usage{
animation.wrapper(time.series, starting.points, location.lim, scale.lim,
  shape.lim, optimization.function = climex:::nmk.modified,
  optimization.steps, width, height, delay = 300, loopMode = "loop",
  image.folder, working.folder)
}
\arguments{
\item{time.series}{Blocked data for which the GEV fit is performed}

\item{starting.points}{of the optimization.}

\item{location.lim}{Range of the location parameter in which the likelihood surface will be calculated.}

\item{scale.lim}{Range of the scale parameter in which the likelihood surface will be calculated.}

\item{shape.lim}{Range of the shape parameter in which the likelihood surface will be calculated.}

\item{optimization.function}{For now only a internal hack of the dfoptim::nmk function is available. This is not really pleasant since it is not the original function of the optimization. I should expand this soon.}

\item{optimization.steps}{Since the first updates contain a lot of jumps and the last ones almost no deviations in the GEV parameters the range of the trajectories can be limited.}

\item{width}{of both the images and the form with the playback options in pixel.}

\item{height}{of both the images and the form with the playback options in pixel.}

\item{delay}{Time for which the individual pictures stay visible during the animation in ms. Default = 300.}

\item{loopMode}{Scianimator parameter declaring what will happen after a full run through all the images. Default = "loop".}

\item{image.folder}{where the generated pictures should be saved.}

\item{working.folder}{the overall folder containing the generated pictures, JavaScript files etc.}
}
\value{
An invisible copy of the generated JavaScript file.
}
\details{
All the images as well as the JavaScript script will be produced in a folder in the /tmp/ directory, while the template will be extracted from the climex app system files. A huge thank towards Brent Ertz for his awesome scianimator jquery plugin \url{https://github.com/brentertz/scianimator} and Yihui Xie who's animation R package was the basis for this code. The warning messages "Removed x rows ..." appear since a couple of trajectories are outside the plotted area.
}
\author{
Philipp Mueller
}
\seealso{
\code{\link{plot.animation}}

Other shiny: \code{\link{climex}},
  \code{\link{plot.animation}}
}

