% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{anomalies}
\alias{anomalies}
\title{Anomalies of a time series}
\usage{
anomalies(x)
}
\arguments{
\item{x}{Time series of the class 'xts'.}
}
\value{
Time series of the class 'xts' of the same length as x containing the anomalies of the series.
}
\description{
Calculates the anomalies of a time series.
}
\details{
Construction via the subtraction of the mean
  (temperature) value of the specific date. I don't really like
  this one. Maybe try to add a weighting function with sharp
  bandwidth instead. Uses the \code{\link[lubridate]{yday}}.
}
\seealso{
Other ts: \code{\link{aic}}, \code{\link{bic}},
  \code{\link{check.completeness}},
  \code{\link{remove.incomplete.years}},
  \code{\link{remove.seasonality}}
}
\author{
Philipp Mueller
}
