% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{anomalies.xts}
\alias{anomalies.xts}
\title{Anomalies of a time series}
\usage{
\method{anomalies}{xts}(x, mc.cores = NULL)
}
\arguments{
\item{x}{A time series of class \pkg{xts}}

\item{mc.cores}{A numerical input specifying the number of cores
to use for the multi core application of the function (see
\code{\link[parallel]{detectCores}}). This functionality is only
available if the input is a list of different objects. If NULL,
the function will be calculated classically with only one core.
Default = NULL.}
}
\value{
Same class as the input
}
\description{
Calculates the anomalies of an object of class
  \pkg{xts} or a list of such objects.
}
\details{
Construction via the subtraction of the mean
  value of the specific date.

  Uses the \code{\link[lubridate]{yday}}.
}
\seealso{
Other ts: \code{\link{aic.climex.fit.gev}},
  \code{\link{aic.climex.fit.gpd}}, \code{\link{aic.list}},
  \code{\link{aic}}, \code{\link{anomalies.list}},
  \code{\link{anomalies}},
  \code{\link{bic.climex.fit.gev}},
  \code{\link{bic.climex.fit.gpd}}, \code{\link{bic.list}},
  \code{\link{bic}}, \code{\link{check.completeness}},
  \code{\link{remove.incomplete.years.list}},
  \code{\link{remove.incomplete.years.xts}},
  \code{\link{remove.incomplete.years}},
  \code{\link{remove.seasonality.list}},
  \code{\link{remove.seasonality.xts}},
  \code{\link{remove.seasonality}}
}
\author{
Philipp Mueller
}
