% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.fevd}
\alias{as.fevd}
\title{Converts fits done by \code{\link{optim}} into an object which can be used by the functions of the extRemes package}
\usage{
as.fevd(x, fit = stats::optim(x), call = "fevd(x = x)", data.name = c("x",
  ""), weights = 1, in.data = FALSE, method = "MLE", type = "GEV",
  period.basis = "year", par.models = list(threshold = ~1, location = ~1,
  scale = ~1, log.scale = FALSE, shape = ~1, term.names = list(threshold =
  character(0), location = character(0), scale = character(0), shape =
  character(0))), const.loc = TRUE, const.scale = TRUE,
  const.shape = TRUE, n = length(x), na.action = "na.fail",
  parnames = c("location", "scale", "shape"), results = fit,
  initial.results = NULL)
}
\arguments{
\item{x}{Time series on which the.}

\item{fit}{GEV fit performed on the data set.}
}
\value{
Of class "fevd"
}
\details{
Caution: the fit performed by the fevd function also calculates the hessian matrix. The element 'initial.results' contains the initial values of the GEV parameters calculated via the heuristic formula of the moments and the lmoments method. But I don't see a point right now to implement this too. In addition 'call' is not a real call.
}
\author{
Philipp Mueller
}

