% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{block.list}
\alias{block.list}
\title{Blocking data}
\usage{
\method{block}{list}(x, block.number = NULL, block.length = NULL,
  block.mode = c("max", "min"), mc.cores = NULL)
}
\arguments{
\item{x}{A list of objects of class \pkg{xts}.}

\item{block.number}{Specifies the number of blocks the input data
is going to be separated in. Default = NULL.}

\item{block.length}{Length of the blocks. For the sake of
simplicity the last block is not forced to match the length of
the other plots. Default = NULL.}

\item{block.mode}{This parameter determines if the maximum "max"
or the minimum "min" of a block shall be returned. Default:
"max".}

\item{mc.cores}{A numerical input specifying the number of cores
to use for the multi core application of the function (see
\code{\link[parallel]{detectCores}}). This functionality is only
available if the input is a list of different objects. If NULL,
the function will be calculated classically with only one core.
Default = NULL.}
}
\value{
Same class as the input.
}
\description{
Separates the input into blocks of equal size and
  returns the maximum or minimum of the block as result.
}
\details{
If both the \strong{block.number} and
  \strong{block.length} arguments are NULL, the data will be
  split into annual blocks. If not, the
  size of a block is determined by the \strong{block.length}
  argument or calculated via the \strong{block.number} argument if
  the previous one is set to NULL. For calculating the mean of
  the blocks, please have a look at the \code{\link[stats]{ave}}
  function.

  This function can be applied to lists of \pkg{xts} class elements
  too.
}
\examples{
  block( temp.potsdam )
}
\seealso{
Other extremes: \code{\link{block.xts}},
  \code{\link{block}}, \code{\link{decluster.list}},
  \code{\link{decluster.xts}}, \code{\link{decluster}},
  \code{\link{extremal.index}}, \code{\link{gev.density}},
  \code{\link{gpd.density}}, \code{\link{qevd}},
  \code{\link{return.level.climex.fit.gev}},
  \code{\link{return.level.list}},
  \code{\link{return.level}}, \code{\link{revd}},
  \code{\link{rlevd}}, \code{\link{threshold.list}},
  \code{\link{threshold.xts}}, \code{\link{threshold}}

Other extremes: \code{\link{block.xts}},
  \code{\link{block}}, \code{\link{decluster.list}},
  \code{\link{decluster.xts}}, \code{\link{decluster}},
  \code{\link{extremal.index}}, \code{\link{gev.density}},
  \code{\link{gpd.density}}, \code{\link{qevd}},
  \code{\link{return.level.climex.fit.gev}},
  \code{\link{return.level.list}},
  \code{\link{return.level}}, \code{\link{revd}},
  \code{\link{rlevd}}, \code{\link{threshold.list}},
  \code{\link{threshold.xts}}, \code{\link{threshold}}
}
\author{
Philipp Mueller
}
