% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{block}
\alias{block}
\title{block}
\usage{
block(input.bulk, block.number = round(nrow(input.bulk)/50),
  block.length = NULL, block.mode = c("max", "min"),
  separation.mode = c("weeks", "months", "years", "quarterly"))
}
\arguments{
\item{input.bulk}{Provided data. Preferably of class 'xts' but also backward compatible.}

\item{block.number}{Specifies the number of blocks the input data is going to be separated in.}

\item{block.length}{Length of the blocks. For the sake of simplicity the last block is not forced to match the length of the other plots.}

\item{block.mode}{This parameter determines if the maximum "max" or the minimum "min" of a block shall be returned. Default: "max".}

\item{separation.mode}{"years" is used to split the data according to its date values instead.}
}
\value{
Same format as iuput.
}
\description{
Separates the input into blocks of equal size and returns the maximum or minimum of the block as result.
}
\details{
If 'separation.mode' is set to "years" the data is separated according to its time stamps. If not the size of a block is determined by the 'block.length' parameter or calculated via the 'block.number' parameter. This is done for data of class 'xts'. For calculating the mean of the blocks have a look at the \code{\link{stats::ave}} function.
}
\examples{
block( temp.potsdam )
}
\author{
Philipp Mueller
}

