% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{check.completeness.list}
\alias{check.completeness.list}
\title{Checking for N complete years of data}
\usage{
\method{check.completeness}{list}(x, number.of.years = 30)
}
\arguments{
\item{x}{A list of \pkg{xts} class objects.}

\item{number.of.years}{The minimum number of complete years in the
series. Default = 30.}
}
\value{
Trimmed list containing only those elements, which indeed
  have more than N years of complete data.
}
\description{
Removes all class \pkg{xts} time series from a
  list, which are not of at least \strong{number.of.years}
  complete years of data.
}
\author{
Philipp Mueller
}
