% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{check.completeness}
\alias{check.completeness}
\title{Checking for N complete years of data}
\usage{
check.completeness(x, number.of.years = 30)
}
\arguments{
\item{x}{Either a time series of class \pkg{xts} or a list of
them.}

\item{number.of.years}{The minimum number of complete years in the
series. Default = 30.}
}
\value{
If the input is of type \pkg{xts}, the function returns
  TRUE or FALSE. If the input, on the other hand, is a list of
  objects of type \pkg{xts}, it returns a trimmed list
  containing only those elements, which indeed have more than N
  years of complete data.
}
\description{
This function tests whether the series do has at least
  \emph{number.of.years} complete years of data.
}
\details{
It uses the \code{\link{remove.incomplete.years}}
  to check for the completeness of the individual years.
}
\seealso{
Other ts: \code{\link{aic}}, \code{\link{anomalies}},
  \code{\link{bic}}, \code{\link{remove.incomplete.years}},
  \code{\link{remove.seasonality}}
}
\author{
Philipp Mueller
}
