% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{check.completeness.xts}
\alias{check.completeness.xts}
\title{Checking for N complete years of data}
\usage{
\method{check.completeness}{xts}(x, number.of.years = 30,
  mc.cores = NULL)
}
\arguments{
\item{x}{A time series of class \pkg{xts}.}

\item{number.of.years}{The minimum number of complete years in the
series. Default = 30.}

\item{mc.cores}{A numerical input specifying the number of cores
to use for the multi core application of the function (see
\code{\link[parallel]{detectCores}}). This functionality is only
available if the input is a list of different objects. If NULL,
the function will be calculated classically with only one core.
Default = NULL.}
}
\value{
Logical
}
\description{
This function tests whether a \pkg{xts} class object
  has at least \strong{number.of.years} complete years of data.
}
\details{
It uses the \code{\link{remove.incomplete.years}}
  to check for the completeness of the individual years.
}
\author{
Philipp Mueller
}
\concept{data.cleaning}
