% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_preprocessing.R
\name{cleaning.interactive}
\alias{cleaning.interactive}
\title{Function to get rid of artifacts within the Climex app}
\usage{
cleaning.interactive(x.xts, checkboxIncompleteYears, checkboxDecluster,
  sliderThreshold)
}
\arguments{
\item{x.xts}{Time series of class 'xts' which has to be cleaned.}

\item{checkboxIncompleteYears}{Logical (checkbox) input determining
whether to remove all incomplete years of a time series. This box
will be only available if input$radioEvdStatistics == "GEV" and else
will be NULL.}

\item{checkboxDecluster}{Logical (checkbox) input determining
whether to remove all clusters in a time series and replace them by
their maximal value. This box will be only available if
input$radioEvdStatistics == "GP" and else will be NULL.}

\item{sliderThreshold}{Numerical (slider) input determining the
threshold used within the GP fit and the extraction of the extreme
events. Boundaries: minimal and maximal value of the deseasonalized
time series (rounded). Default: 0.8* the upper end point. This one
is only used in declustering the time series.}
}
\value{
Time series of class 'xts'.
}
\details{
The app does two things: First it replaces all -999 and NaN
in the time series by NaN (the former is the default indicator in
the data of the German weather service for missing values).
Second it removes all incomplete years (GEV) or cluster (GP) when
the corresponding checkbox is checked.
}
\seealso{
Other preprocessing: \code{\link{data.extremes}},
  \code{\link{deseasonalize.interactive}},
  \code{\link{deseasonalizeSelectionInput}},
  \code{\link{deseasonalizeSelection}},
  \code{\link{extremes.interactive}},
  \code{\link{generalExtremeExtractionInput}},
  \code{\link{generalExtremeExtraction}}
}
\author{
Philipp Mueller
}
