% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{climex}
\alias{climex}
\title{climex}
\usage{
climex(x.input = NULL)
}
\arguments{
\item{Time}{series of type xts or list of such time series. Default = NULL}
}
\value{
starts a shiny app.
}
\description{
Shiny app combining most of the tools used in extreme value analysis using GEV fitting via maximum likelihood
}
\details{
It possible to provide a time series of type xts to a list of elements of this type. This app need the its own css file.
}
\author{
Philipp Mueller
}
\seealso{
Other shiny: \code{\link{likelihood.gui}}
}

