% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_base.R
\name{climex.server}
\alias{climex.server}
\title{Server-side part of the \code{\link{climex}} function.}
\usage{
climex.server(input, output)
}
\value{
Function acting as the shiny server.
}
\details{
Since it grew organically most of the features are defined inside the function
}
\author{
Philipp Mueller
}

