% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_base.R
\name{climex.ui}
\alias{climex.ui}
\title{The user interface for the \code{\link{climex}} function.}
\usage{
climex.ui()
}
\value{
HTML code of the interface
}
\details{
Contains all the HTML codes.
}
\author{
Philipp Mueller
}

