% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_years.R
\name{collect_years}
\alias{collect_years}
\title{Title}
\usage{
collect_years(start, end, arl)
}
\arguments{
\item{start}{vector of timesteps indicating the start of each year}

\item{end}{vector of timesteps indicating the end of each year}

\item{arl}{list of arrays containing the data where each list entry is a single timestep}
}
\value{
a list of arrays where each list entry is a single year
}
\description{
Title
}
