% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{color.table}
\alias{color.table}
\title{Adds font color to a HTML table.}
\usage{
color.table(x.html.table, css.colours,
  style = "table-condensed table-bordered")
}
\arguments{
\item{x.html.table}{HTML table.}

\item{css.colours}{Character vector containing the colors in hex.}

\item{style}{Additional style tags for the output table.}
}
\value{
Same format as input.
}
\details{
In each element of the table where the font should be colored, the corresponding color has to be added via a paste( as.character( table[ x, y ] ), "#0000FF" ).
}
\author{
Philipp Mueller
}
\seealso{
\code{\link{pander::pandoc.table.return}} and \code{\link{markdown::markdownToHTML}}
}

