% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert.date}
\alias{convert.date}
\title{Convert data into dates and back.}
\usage{
convert.date(input.data, origin = "1970-01-01")
}
\arguments{
\item{input.data}{Data containing a date entry in its first column.}

\item{origin}{Sets origin date for the transformation of numerical
data into class \emph{Date}. Per default its set to "1970-01-01"
like its the convention in many software including R.}
}
\value{
Returns input in same format with second column of class
  date
}
\description{
Converts class \emph{numeric} into \emph{Date} and
  back and also transforms the DWD input (class \emph{integer})
  into class \emph{Date}.
}
\details{
The date entry has to be provided in the first column.
  Generic function for the following input classes:
  \emph{integer}, \emph{numeric}, \emph{character}, and
  \emph{Date}.  For an easy way to produce a specific date a
  string ala "20140101" can be provided.
}
\author{
Philipp Mueller
}
