% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data}
\alias{create_data}
\title{Title}
\usage{
create_data(year_list)
}
\arguments{
\item{year_list}{list of arrays for each year}
}
\value{
a single array containing the finished data to be written to netcdf
}
\description{
Title
}
