% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_leaflet.R
\name{data.chosen}
\alias{data.chosen}
\title{Choosing a data set in the Climex app}
\usage{
data.chosen(selectDataBase, sliderYears, selectDataType, reactive.loading,
  climex.environment)
}
\arguments{
\item{selectDataBase}{Character (select) input to determine the data
source. In the default installation there are three options:
c( "Input", "DWD", "Artificial data" ). The first one uses the data
provided as an argument to the call of the \code{\link{climex}}
function. The second one uses the database of the German weather
service (see \code{\link{download.data.dwd}}). The third one allows
the user to produce random numbers distributed according to the GEV
or GP distribution. Determined by menuSelectDataBase.
Default = "DWD".}

\item{sliderYears}{Numerical (slider) input to determine the minimal
length (in years) of the time series to be displayed. Minimal value
is 0 and maximal is 155 (longest one in the DWD database), the
default value is 65 and the step width is 1.}

\item{selectDataType}{Character (select) input to determine which set
measurements should be used for the selected station. In case of the
default import of the DWD data, there are three options:
c( "Daily max. temp", "Daily min. temp", "Daily precipitation" ).
Determined by menuSelectDataType.}

\item{reactive.loading}{Reactive value allowing the user to load a
time series of class "xts" or "list" with "xts" as their elements
into the climex app. \code{\link{file.loading}}}

\item{climex.environment}{Environment containing the global
variables used within the climex app. Namely the last values
displayed in the table and the lists containing the station
data.}
}
\value{
Reactive list containing a list of all selected stations and
their positions.
}
\description{
This functions extracts all stations containing more
  than a specified number of years of data
}
\details{
It uses the current database and returns all stations which
are at least as long as the value of the input$sliderYears slider.
}
\seealso{
Other leaflet: \code{\link{leafletClimexUI}},
  \code{\link{leafletClimex}}
}
\author{
Philipp Mueller
}
