% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_fitting.R
\name{data.initials}
\alias{data.initials}
\title{Reactive value containing the initial parameters to start the
GEV/GP fit at.}
\usage{
data.initials(initialLocation, initialScale, initialShape, radioEvdStatistics,
  reactive.extreme)
}
\arguments{
\item{initialLocation}{Numerical (numerical) input determining the
location parameter at which to start the fitting procedure.}

\item{initialScale}{Numerical (numerical) input determining the
scale parameter at which to start the fitting procedure.}

\item{initialShape}{Numerical (numerical) input determining the
shape parameter at which to start the fitting procedure.}

\item{radioEvdStatistics}{Character (radio) input determining whether
the GEV or GP distribution shall be fitted to the data. Choices:
c( "GEV", "GP" ), default = "GEV".}

\item{reactive.extreme}{Reactive value returning a list containing
three elements: 1. the blocked time series, 2. the deseasonalized time
series, and 3. the pure time series.}
}
\value{
Numerical vector containing three elements: location, scale
and shape parameter to start the GEV fit at.
}
\details{
The initial parameter combination can be set by the numerical
input in the Likelihood tab. Until that happened, just the default
values obtained by \code{\link{likelihood.initials}} will be used.
}
\author{
Philipp Mueller
}
\seealso{
Other climex-fitting: \code{\link{data.fitting}},
  \code{\link{fit.interactive}},
  \code{\link{generalFittingRoutineInput}}
}

