% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_preprocessing.R
\name{deseasonalize.interactive}
\alias{deseasonalize.interactive}
\title{Function for removing the seasonality of a given time series
within the Climex app.}
\usage{
deseasonalize.interactive(x.xts, selectDeseasonalize, selectDataBase)
}
\arguments{
\item{x.xts}{Time series of class 'xts' which has to be cleaned.}

\item{selectDeseasonalize}{Character (select) input determining which
deseasonalization method should be used to remove the short-range
correlations from the provided time series.
\code{\link{deseasonalizeInput}}. If NULL climex::anomalies will be
used.}

\item{selectDataBase}{Character (select) input to determine the data
source. In the default installation there are three options:
c( "Input", "DWD", "Artificial data" ). The first one uses the data
provided as an argument to the call of the \code{\link{climex}}
function. The second one uses the database of the German weather
service (see \code{link{download.data.dwd}}). The third one allows
the user to produce random numbers distributed according to the GEV
or GP distribution. Determined by menuSelectDataBase.
Default = "DWD".}
}
\value{
Time series of class 'xts'.
}
\seealso{
Other preprocessing: \code{\link{cleaning.interactive}},
  \code{\link{data.extremes}},
  \code{\link{deseasonalizeSelectionInput}},
  \code{\link{deseasonalizeSelection}},
  \code{\link{extremes.interactive}},
  \code{\link{generalButtonMinMaxInput}},
  \code{\link{generalButtonMinMax}},
  \code{\link{generalExtremeExtractionInput}},
  \code{\link{generalExtremeExtraction}}
}
\author{
Philipp Mueller
}
