% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_preprocessing.R
\name{deseasonalizeInput}
\alias{deseasonalizeInput}
\title{Removing the seasonality.}
\usage{
deseasonalizeInput()
}
\value{
selectInput
}
\details{
Not a real shiny module, since I have to use this select
input outside its namespace.
}
\author{
Philipp Mueller
}
\seealso{
Other preprocessing: \code{\link{cleaning.interactive}},
  \code{\link{data.extremes}},
  \code{\link{deseasonalize.interactive}},
  \code{\link{extremes.interactive}},
  \code{\link{generalExtremeExtractionInput}},
  \code{\link{generalExtremeExtraction}}
}

