% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_preprocessing.R
\name{deseasonalizeSelection}
\alias{deseasonalizeSelection}
\title{Removing the seasonality in the Climex app}
\usage{
deseasonalizeSelection(selectDataBase)
}
\arguments{
\item{selectDataBase}{Character (select) input to determine the data
source. In the default installation there are three options:
c( "Input", "DWD", "Artificial data" ). The first one uses the data
provided as an argument to the call of the \code{\link{climex}}
function. The second one uses the database of the German weather
service (see \code{\link{download.data.dwd}}). The third one allows
the user to produce random numbers distributed according to the GEV
or GP distribution. Determined by menuSelectDataBase.
Default = "DWD".}
}
\value{
selectInput
}
\description{
Removing the seasonality.
}
\details{
Not a real shiny module, since I have to use this select
input outside its namespace.
}
\seealso{
Other preprocessing: \code{\link{cleaning.interactive}},
  \code{\link{data.extremes}},
  \code{\link{deseasonalize.interactive}},
  \code{\link{deseasonalizeSelectionInput}},
  \code{\link{extremes.interactive}},
  \code{\link{generalButtonMinMaxInput}},
  \code{\link{generalButtonMinMax}},
  \code{\link{generalExtremeExtractionInput}},
  \code{\link{generalExtremeExtraction}}
}
\author{
Philipp Mueller
}
