% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{download.data.dwd}
\alias{download.data.dwd}
\title{Downloads data of the DWD}
\usage{
download.data.dwd(save.downloads = TRUE, csv.export = FALSE,
  data.type = c("default", "temp.max", "temp.min", "prec", "temp.mean",
  "vapor.pressure", "cloud.amount", "air.pressure", "relative.humidity",
  "wind.speed", "temp.min.at.ground", "wind.speed.peak", "prec.type", "quality",
  "sunshine.duration", "snow.height"), download.path = NULL)
}
\arguments{
\item{save.downloads}{If TRUE the downloaded .zip files are stored
in download.path/downloads_dwd. Else they will be deleted after the
extracting. Default = TRUE.}

\item{csv.export}{If TRUE creates an additional folder containing
.csv files with the individual station data. Using this the data can
be used outside of R too. Default = FALSE.}

\item{data.type}{Specifies which kind of information from the
downloaded files should be extracted. This input can be a character
vector  The options are: temp.max, temp.min, prec, default (for both
the daily maximum and minimum temperature and the precipitation),
temp.mean, vapor.pressure, cloud.amount, air.pressure,
quality (-999 = faulty or suspicious, 1 = superficially checked,
2 = checked according to individual criterion, 3 = old automated
check and correction, 5 = historical and subjective check, 7 =
additional second check but without correction, 8 = quality check
outside of the routine (?, just quoting the manual), 9 = not all
parameters are corrected, 10 = fully checked and corrected),
relative.humidity, wind.speed, temp.min.at.ground, wind.speed.peak,
prec.type (0 = no precipitation, 1 = only rain (before 1979), 2 =
unknown, 4 = only rain (after 1979), 7 = only snow, 8 = snow or rain),
sunshine.duration, snow.height. Default = default.}

\item{download.path}{Specifies the data will be stored and downloaded
too. It is advised to store it in the path stored in the options(
"climex.path" ),  which is also used for importing the saved data. You
can overwrite its default value of "~/R/climex/" by adding
options( climex.path = "PATH" ) to your .Rprofile path in your home.}
}
\value{
invisible setwd()
}
\description{
Downloads daily weather data from observation
  stations in Germany and extracts minimum and maximum temperature
  as well as precipitation data.
}
\details{
The download will be done using 'wget'. Per default the
climex.path variable from the getOption( "climex.path" ) will be
used to set the download path. Since this function will check the
files already present it's strongly recommended to use the
save.downloads options. Whenever this function is invoked again only
updated files will be downloaded which saves a lot of traffic and
time. The csv.export option can be used to export the time series
into a data type file making it available outside of R too. In
addition the geographic positions of the individual stations will
be extracted and saved as well. They are needed for the leaflet
module of the climex shiny app. CAUTION: since this procedure takes
a while its run in parallel on all cores of your machine!
}
\author{
Philipp Mueller
}
