% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_climate.R
\name{extend_climate}
\alias{extend_climate}
\title{Title}
\usage{
extend_climate(
  nc_file,
  var_name,
  out_file,
  ext_period = 100,
  origin = "months since 2100-01-01 12:00:00"
)
}
\arguments{
\item{nc_file}{file for which data should be extended}

\item{var_name}{variable name to use}

\item{out_file}{file name for the resulting extended data}

\item{ext_period}{how long the data should be extended for}

\item{origin}{cf compliant unit for time dimension i.e. "days since 1979-01-01 12:00:00"}
}
\value{
a netcdf file
}
\description{
Title
}
