% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{extremal.index}
\alias{extremal.index}
\title{Calculates the extremal index of a time series}
\usage{
extremal.index(x, threshold, method = c("intervals", "runs"),
  minimal.distance = 3)
}
\arguments{
\item{x}{Time series of class 'xts' or numerical vector.}

\item{threshold}{Only events exceeding a specific threshold will be
considered extreme event and thus will be subject of the
declustering.}

\item{method}{"runs" considers a event to start a new cluster only
after run.length events already were below the supplied threshold
(see Stuart Coles). "intervals" is doing something different.
Default = "intervals".}

\item{minimal.distance}{Number of events below the threshold until
the following event will be considered the starting point of a new
cluster. This only affects the "runs" method. Default = 3.}
}
\value{
Numerical vector containing c( extremal index, number of
clusters, minimal distance between clusters (minimal.distance) ).
}
\details{
Started as a reduced version of the extRemes::extremalindex
function. The extremal index can be thought of as the inverse of the
mean cluster size. It can be calculated either by the "runs" method
described by Stuart Coles, 2001 p. 100 or by the "intervals" method
of Ferro and Segers, 2003.
}
\author{
Philipp Mueller
}
\seealso{
Other extremes: \code{\link{block}},
  \code{\link{decluster}}, \code{\link{gev.density}},
  \code{\link{gpd.density}}, \code{\link{qevd}},
  \code{\link{return.level}}, \code{\link{revd}},
  \code{\link{rlevd}}, \code{\link{threshold}}
}

