% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_preprocessing.R
\name{extremes.interactive}
\alias{extremes.interactive}
\title{Function to extract the extreme event from a time series.}
\usage{
extremes.interactive(x.xts, buttonMinMax, radioEvdStatistics, sliderBlockLength,
  sliderThreshold, checkBoxDecluster)
}
\arguments{
\item{x.xts}{Time series of class 'xts' which has to be cleaned.}

\item{buttonMinMax}{Character (radio) input determining whether
the GEV/GP distribution shall be fitted to the smallest or biggest
vales. Choices: c( "Max", "Min ), default = "Max".}

\item{radioEvdStatistics}{Character (radio) input determining whether
the GEV or GP distribution shall be fitted to the data. Choices:
c( "GEV", "GP" ), default = "GEV".}

\item{sliderBlockLength}{Numerical (slider) input determining the
block length used in the GEV flavor of extreme value theory. On
default it is set to one year.}

\item{sliderThreshold}{Numerical (slider) input determining the
threshold used within the GP fit and the extraction of the extreme
events. Boundaries: minimal and maximal value of the deseasonalized
time series (rounded). Default: 0.8* the upper end point.}

\item{checkBoxDecluster}{Logical (checkbox) input determining
whether to remove all clusters in a time series and replace them by
their maximal value. This box will be only available if
input$radioEvdStatistics == "GP" and else will be NULL.}
}
\value{
Time series of class 'xts'.
}
\details{
If the input$radioEvdStatistics is set to "GEV" the time
series will be block. If it's on the other hand set to "GP", all
values above a certain threshold will be extracted.
}
\author{
Philipp Mueller
}
\seealso{
Other preprocessing: \code{\link{cleaning.interactive}},
  \code{\link{deseasonalize.interactive}},
  \code{\link{deseasonalizeInput}},
  \code{\link{generalExtremeExtractionInput}},
  \code{\link{generalExtremeExtraction}}
}

