% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{file.loading}
\alias{file.loading}
\title{File loading in the Climex app}
\usage{
file.loading(fileInputSelection)
}
\arguments{
\item{fileInputSelection}{File input through a shiny base selection
window.}
}
\value{
Reactive value containing either an xts class time series,
a list of those time series or a list of those lists and an
additional data.frame containing meta information about all of the
named stations in those list.
}
\description{
Reactive value listening for a file input via the
  sidebar.
}
\details{
To use this function input$selectDataBase has to be set to
"Input".
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarImprintInput}},
  \code{\link{sidebarImprint}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}},
  \code{\link{sidebarSeriesLengthInput}},
  \code{\link{sidebarSeriesLength}}
}
\author{
Philipp Mueller
}
