% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{find.block.length}
\alias{find.block.length}
\title{find.block.length}
\usage{
find.block.length(x, size.low = 20, size.high = 400, return.period = 100,
  plot = TRUE, main = NULL)
}
\arguments{
\item{x}{Time series of class "xts" or "ts"}

\item{size.low}{Smallest block length applied to the data. Default = 10.}

\item{size.high}{Blggest block length applied to the data. Default = 400.}

\item{return.period}{Argument of the extRemes::return.level() function specifying the return period of the event the return level shall be calculated for. Default = 100.}

\item{plot}{Plots the results. Default = TRUE}

\item{main}{Title of the plot.}

\item{...}{Additional arguments applied for extRemes::fevd()}
}
\value{
Data.frame containing the summary statistic of the analysis:
\itemize{
 \item{ block = specific block length. }
 \item{ loc = MLE estimate for the mean location parameter }
 \item{ se.loc = MLE estimate for the standard deviation of the location parameter. }
 \item{ sca= MLE estimate for the mean scale parameter }
 \item{ se.sca = MLE estimate for the standard deviation of the scale parameter. }
 \item{ sha = MLE estimate for the mean shape parameter }
 \item{ se.sha = MLE estimate for the standard deviation of the shape parameter. }
 \item{ nllh = Negative log-likelihood of the applied fit. }
 \item{ AIC = Akaike information criterion of the applied fit. }
 \item{ BIC = Bayesian information criterion of the applied fit. }
}
}
\description{
Finding the appropriate block length before applying GEV fitting.
}
\details{
Function fits GEV using gev.fit() after applying various block lengths and returns summary statistics. After a sufficient high block length the parameters of the distribution should converge. Function is implemented for inputs of the classes "xts" and "ts".
}
\author{
Philipp Mueller
}

