% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_fitting.R
\name{fit.interactive}
\alias{fit.interactive}
\title{Function to perform the GEV/GP fit within the Climex app.}
\usage{
fit.interactive(x.kept, x.initial = NULL, radioEvdStatistics,
  selectOptimization, buttonMinMax, checkBoxRerun, sliderThreshold)
}
\arguments{
\item{x.kept}{Time series of class 'xts'. Removing clicked or brushed
values has to be done beforehand.}

\item{x.initial}{Initial parameters to start the fitting routine at. See
\code{\link{likelihood.initials}}. Default = NULL.}

\item{radioEvdStatistics}{Character (radio) input determining whether
the GEV or GP distribution shall be fitted to the data. Choices:
c( "GEV", "GP" ), default = "GEV".}

\item{selectOptimization}{Character (select) input to determine which
optimization routine/method is going to be used when fitting the maximum
likelihood function of the GEV/GP distribution. The choices are given in
\code{\link{generalFittingRoutineInput}} and the default value is set to
"Nelder-Mead".}

\item{buttonMinMax}{Character (radio) input determining whether
the GEV/GP distribution shall be fitted to the smallest or biggest
vales. Choices: c( "Max", "Min ), default = "Max".}

\item{checkBoxRerun}{Logical (checkbox) input from the Likelihood tab.
It determines whether or not to start the optimization at the results
of the first run again to escape local minima.}

\item{sliderThreshold}{Numerical (slider) input determining the
threshold used within the GP fit and the extraction of the extreme
events. Boundaries: minimal and maximal value of the deseasonalized
time series (rounded). Default: 0.8* the upper end point.}
}
\value{
Object of class 'climex.fit.gev' or 'climex.fit.gpd', depending
on the choice in input$radioEvdStatistics.
}
\details{
This function does not wait for the initialization of its
slider, checkbox etc. inputs. This way the fit can be performed with its
default settings in the leaflet tab without switching to the General tab
first.
}
\author{
Philipp Mueller
}
\seealso{
Other climex-fitting: \code{\link{color.table}},
  \code{\link{data.fitting}}, \code{\link{data.initials}},
  \code{\link{generalFitStatisticsTable}},
  \code{\link{generalFitStatistics}},
  \code{\link{generalFittingRoutineInput}}
}

