% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_plotting.R
\name{function.get.y.label}
\alias{function.get.y.label}
\title{Determines the propoper label of the y label of the plots in
the General tab of the Climex app.}
\usage{
function.get.y.label(selectDataBase, selectDataType)
}
\arguments{
\item{selectDataBase}{Character (select) input to determine the data
source. In the default installation there are three options:
c( "input", "DWD", "artificial data" ). The first one uses the data
provided as an argument to the call of the \code{\link{climex}}
function. The second one uses the database of the German weather
service (see \code{link{download.data.dwd}}). The third one allows
the user to produce random numbers distributed according to the GEV
or GP distribution. Determined by menuSelectDataBase.
Default = "DWD".}

\item{selectDataType}{Character (select) input to determine which set
measurements should be used for the selected station. In case of the
default import of the DWD data, there are three options:
c( "Daily max. temp", "Daily min. temp", "Daily precipitation" ).
Determined by menuSelectDataType.}
}
\value{
Character
}
\details{
It's more or less just the input$selectDataType being
considered. For artificial data the label will be "temperature".
}
\author{
Philipp Mueller
}

