% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_preprocessing.R
\name{generalButtonMinMaxInput}
\alias{generalButtonMinMaxInput}
\title{Whether to determine the minimal or maximal extremes.}
\usage{
generalButtonMinMaxInput()
}
\value{
menuItemOutput
}
\details{
Not a real shiny module, since I have to use this select
input outside its namespace. Provides the
shinydashboard::menuItemOutput for \code{\link{
generalButtonMinMaxInput}}
}
\seealso{
Other preprocessing: \code{\link{cleaning.interactive}},
  \code{\link{data.extremes}},
  \code{\link{deseasonalize.interactive}},
  \code{\link{deseasonalizeSelectionInput}},
  \code{\link{deseasonalizeSelection}},
  \code{\link{extremes.interactive}},
  \code{\link{generalButtonMinMax}},
  \code{\link{generalExtremeExtractionInput}},
  \code{\link{generalExtremeExtraction}}
}
\author{
Philipp Mueller
}
