% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_preprocessing.R
\name{generalExtremeExtraction}
\alias{generalExtremeExtraction}
\title{Extract the extreme events from a given time series.}
\usage{
generalExtremeExtraction(radioEvdStatistics, deseasonalize.interactive,
  buttonMinMax, reactive.selection)
}
\arguments{
\item{radioEvdStatistics}{Character (radio) input determining whether
the GEV or GP distribution shall be fitted to the data. Choices:
c( "GEV", "GP" ), default = "GEV".}

\item{deseasonalize.interactive}{Function used to remove seasonality
from a given time series.}

\item{buttonMinMax}{Character (radio) input determining whether
the GEV/GP distribution shall be fitted to the smallest or biggest
vales. Choices: c( "Max", "Min ), default = "Max".}

\item{reactive.selection}{Reactive value contains the xts type time
series of the individual station/input chosen via the sidebar or the
leaflet map.}
}
\value{
renderMenu
}
\details{
Provides a slider input to determine either the block length
(in case of the GEV distribution) or the height of the threshold (GP)
}
\author{
Philipp Mueller
}
\seealso{
Other preprocessing: \code{\link{cleaning.interactive}},
  \code{\link{generalExtremeExtractionInput}}
}

