% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_preprocessing.R
\name{generalExtremeExtractionInput}
\alias{generalExtremeExtractionInput}
\title{Extracting extreme events in the Climex app}
\usage{
generalExtremeExtractionInput()
}
\value{
menuItemOutput
}
\description{
Extract the extreme events from a given time series.
}
\details{
Provides the \code{\link[shinydashboard]{menuItemOutput}}
  for \code{\link{generalExtremeExtraction}}. See the later one
  for details.
}
\seealso{
Other preprocessing: \code{\link{cleaning.interactive}},
  \code{\link{data.extremes}},
  \code{\link{deseasonalize.interactive}},
  \code{\link{deseasonalizeSelectionInput}},
  \code{\link{deseasonalizeSelection}},
  \code{\link{extremes.interactive}},
  \code{\link{generalButtonMinMaxInput}},
  \code{\link{generalButtonMinMax}},
  \code{\link{generalExtremeExtraction}}
}
\author{
Philipp Mueller
}
