% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_plotting.R
\name{generalFitPlot}
\alias{generalFitPlot}
\title{Plot of the fits in the Climex app}
\usage{
generalFitPlot(input, output, session, reactive.extreme, reactive.rows,
  reactive.fitting, buttonMinMax, radioEvdStatistics, function.get.y.label,
  selectDataBase, selectDataType, sliderThreshold)
}
\arguments{
\item{input}{Namespace input. For more details check out
\url{http://shiny.rstudio.com/articles/modules.html}}

\item{output}{Namespace output.}

\item{session}{Namespace session.}

\item{reactive.extreme}{Reactive value containing a list of the
extracted extreme events of a time series, the deseasonalized and the
pure version of this very time series. All three elements are
  provided as class 'xts'.}

\item{reactive.rows}{Reactive value holding a logical vector
indicating which values of the time series provided by
\code{\link{data.extremes}} to use after clicking and brushing.}

\item{reactive.fitting}{Reactive value containing the results of the
fit (\code{\link{fit.gev}} or \code{\link{fit.gpd}} depending on
radioEvdStatistic) to the blocked time series in
reactive.extreme()[[ 1 ]].}

\item{buttonMinMax}{Character (radio) input determining whether
the GEV/GP distribution shall be fitted to the smallest or biggest
vales. Choices: c( "Max", "Min ), default = "Max".}

\item{radioEvdStatistics}{Character (radio) input determining whether
the GEV or GP distribution shall be fitted to the data. Choices:
c( "GEV", "GP" ), default = "GEV".}

\item{function.get.y.label}{Function determining the label of the y axis
in all three plots. \code{\link{function.get.y.label}}. You might wonder
why I am not just referring to the package internal function right away
instead of requiring it as an input. Well, this way the whole app
becomes way more easy to customize.}

\item{selectDataBase}{Character (select) input to determine the data
source. In the default installation there are three options:
c( "Input", "DWD", "Artificial data" ). The first one uses the data
provided as an argument to the call of the \code{\link{climex}}
function. The second one uses the database of the German weather
service (see \code{\link{download.data.dwd}}). The third one allows
the user to produce random numbers distributed according to the GEV
or GP distribution. Determined by menuSelectDataBase.
Default = "DWD".}

\item{selectDataType}{Character (select) input to determine which set
measurements should be used for the selected station. In case of the
default import of the DWD data, there are three options:
c( "Daily max. temp", "Daily min. temp", "Daily precipitation" ).
Determined by menuSelectDataType.}

\item{sliderThreshold}{Numerical (slider) input determining the
threshold used within the GP fit and the extraction of the extreme
events. Boundaries: minimal and maximal value of the deseasonalized
time series (rounded). Default: 0.8* the upper end point.}
}
\value{
Nothing in particular.
}
\description{
Box containing the plot of the performed GEV/GP
  fitting as well as three goodness-of-fit plots.
}
\details{
All four plots are rendered using the ggplot2 package.
}
\seealso{
Other climex-plot: \code{\link{function.get.y.label}},
  \code{\link{generalFitPlotOutput}},
  \code{\link{generalTimeSeriesPlotOutput}},
  \code{\link{generalTimeSeriesPlot}}
}
\author{
Philipp Mueller
}
