% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_plotting.R
\name{generalFitPlotOutput}
\alias{generalFitPlotOutput}
\title{Plot of the fits in the Climex app}
\usage{
generalFitPlotOutput(id)
}
\arguments{
\item{id}{Namespace prefix}
}
\value{
tabBox
}
\description{
Box containing the plot of the performed GEV/GP
  fitting as well as three goodness-of-fit plots.
}
\details{
This function provides the ouput to
\code{\link{generalFitPlot}}. All four plots are rendered using the
ggplot2 package.
}
\seealso{
Other climex-plot: \code{\link{function.get.y.label}},
  \code{\link{generalFitPlot}},
  \code{\link{generalTimeSeriesPlotOutput}},
  \code{\link{generalTimeSeriesPlot}}
}
\author{
Philipp Mueller
}
