% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_fitting.R
\name{generalFitStatistics}
\alias{generalFitStatistics}
\title{Result table in Climex app}
\usage{
generalFitStatistics(reactive.fitting, reactive.extreme, sliderThreshold,
  buttonMinMax, radioEvdStatistics, color.table, climex.environment)
}
\arguments{
\item{reactive.fitting}{Reactive value containing the results of the
fit (\code{\link{fit.gev}} or \code{\link{fit.gpd}} depending on
radioEvdStatistic) to the blocked time series in
reactive.extreme()[[ 1 ]].}

\item{reactive.extreme}{Reactive value containing a list of the
extracted extreme events of a time series, the deseasonalized and the
pure version of this very time series. All three elements are provided
as class 'xts'.}

\item{sliderThreshold}{Numerical (slider) input determining the
threshold used within the GP fit and the extraction of the extreme
events. Boundaries: minimal and maximal value of the deseasonalized
time series (rounded). Default: 0.8* the upper end point.}

\item{buttonMinMax}{Character (radio) input determining whether
the GEV/GP distribution shall be fitted to the smallest or biggest
vales. Choices: c( "Max", "Min ), default = "Max".}

\item{radioEvdStatistics}{Character (radio) input determining whether
the GEV or GP distribution shall be fitted to the data. Choices:
c( "GEV", "GP" ), default = "GEV".}

\item{color.table}{Function adding color to the table constructed
within generalFitStatistics. It will replace some placeholders by
color tags. \code{\link{color.table}}}

\item{climex.environment}{Environment containing the global
variables used within the climex app. Namely the last values
displayed in the table and the lists containing the station
data.}
}
\value{
renderUI containing a table
}
\description{
Table to display the results of the GEV/GP fitting
  procedure.
}
\details{
Displaying of AIC, nllh, BIC and fitted parameters as
  well as the difference to the three last fits! (and highlight
  positive values with with green and negative with red). This
  function will define some global variables to hold the results
  of the former fits. This is necessary in order to mark the
  progress in green or red.
}
\seealso{
Other climex-fitting: \code{\link{color.table}},
  \code{\link{data.fitting}},
  \code{\link{fit.interactive}},
  \code{\link{generalFitStatisticsTable}}
}
\author{
Philipp Mueller
}
