% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_fitting.R
\name{generalFitStatisticsTable}
\alias{generalFitStatisticsTable}
\title{Table to display the results of the GEV/GP fitting procedure.}
\usage{
generalFitStatisticsTable()
}
\value{
box
}
\details{
Provides the UI part of \code{\link{generalFitStatistics}}
but is not a proper Shiny module.
}
\author{
Philipp Mueller
}
\seealso{
Other climex-fitting: \code{\link{color.table}},
  \code{\link{data.fitting}}, \code{\link{data.initials}},
  \code{\link{fit.interactive}},
  \code{\link{generalFitStatistics}},
  \code{\link{generalFittingRoutineInput}}
}

