% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_fitting.R
\name{generalFittingRoutineInput}
\alias{generalFittingRoutineInput}
\title{Selection input to determine the fitting routine/method}
\usage{
generalFittingRoutineInput()
}
\value{
selectInput
}
\author{
Philipp Mueller
}
\seealso{
Other climex-fitting: \code{\link{fit.interactive}}
}

