% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_plotting.R
\name{generalTimeSeriesPlot}
\alias{generalTimeSeriesPlot}
\title{TabBox containing a plot of the pure and deseasonalized time
series as well as one of all its extreme events.}
\usage{
generalTimeSeriesPlot(input, output, session, reactive.extreme, selectDataBase,
  selectDataType, function.get.y.label, radioEvdStatistics, sliderThreshold)
}
\arguments{
\item{input}{Namespace input. For more details check out
\link{ \url{ http://shiny.rstudio.com/articles/modules.html } }}

\item{output}{Namespace output.}

\item{session}{Namespace session.}

\item{reactive.extreme}{Reactive value containing a list of the
extracted extreme events of a time series, the deseasonalized and the
pure version of this very time series. All three elements are provided
as class 'xts'.}

\item{selectDataBase}{Character (select) input to determine the data
source. In the default installation there are three options:
c( "input", "DWD", "artificial data" ). The first one uses the data
provided as an argument to the call of the \code{\link{climex}}
function. The second one uses the database of the German weather
service (see \code{link{download.data.dwd}}). The third one allows
the user to produce random numbers distributed according to the GEV
or GP distribution. Determined by menuSelectDataBase.
Default = "DWD".}

\item{selectDataType}{Character (select) input to determine which set
measurements should be used for the selected station. In case of the
default import of the DWD data, there are three options:
c( "Daily max. temp", "Daily min. temp", "Daily precipitation" ).
Determined by menuSelectDataType.}

\item{function.get.y.label}{Function determining the label of the y axis
in all three plots. \code{\link{function.get.y.label}}. You might wonder
why I am not just referring to the package internal function right away
instead of requiring it as an input. Well, this way the whole app
becomes way more easy to customize.}

\item{radioEvdStatistics}{Character (radio) input determining whether
the GEV or GP distribution shall be fitted to the data. Choices:
c( "GEV", "GP" ), default = "GEV".}

\item{sliderThreshold}{Numerical (slider) input determining the
threshold used within the GP fit and the extraction of the extreme
events. Boundaries: minimal and maximal value of the deseasonalized
time series (rounded). Default: 0.8* the upper end point.}
}
\value{
Reactive value holding a logical vector indicating which values
of the time series provided by \code{\link{data.extremes}} to use after
clicking and brushing.
}
\details{
Both the pure time series and the
deseasonalized one are rendered using the dygraphs package. The extreme
events are rendered using the ggplot2 package. For the later one
clicking or brushing points enables the user to exclude them. This is
handled by a reactive value returned by this function.
}
\author{
Philipp Mueller
}
\seealso{
Other climex-plot: \code{\link{function.get.y.label}},
  \code{\link{generalFitPlotOutput}},
  \code{\link{generalFitPlot}},
  \code{\link{generalTimeSeriesPlotOutput}}
}

