% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_plotting.R
\name{generalTimeSeriesPlotOutput}
\alias{generalTimeSeriesPlotOutput}
\title{TabBox containing a plot of the pure and deseasonalized time
series as well as one of all its extreme events.}
\usage{
generalTimeSeriesPlotOutput(id)
}
\arguments{
\item{id}{Namespace prefix}
}
\value{
tabBox
}
\details{
This function provides the ouput to
\code{\link{generalTimeSeriesPlot}}. Both the pure time series and the
deseasonalized one are rendered using the dygraphs package. The extreme
events are rendered using the ggplot2 package.
}
\author{
Philipp Mueller
}
\seealso{
Other climex-plot: \code{\link{function.get.y.label}},
  \code{\link{generalTimeSeriesPlot}}
}

