% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{get_template}
\alias{get_template}
\title{Title}
\usage{
get_template(nc, ext_period, origin)
}
\arguments{
\item{nc}{the netcdf file read in with nc_open}

\item{ext_period}{integer, years to extend data by}

\item{origin}{cf compliant unit for time dimension i.e. "days since 1979-01-01 12:00:00"}
}
\value{
a variable template to create a new netcdf file with the same structure
}
\description{
Title
}
