% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{gev.density}
\alias{gev.density}
\title{Port of the ismev::gev.dens function}
\usage{
gev.density(parameters, z)
}
\arguments{
\item{parameters}{Fitted GEV parameters c( location, scale,
shape )}

\item{z}{Numerical vector of sites where to evaluate the
density of the GEV distribution.}
}
\value{
Numerical vector of same length as z.
}
\details{
It might not to be called by the user, but I can
certainly write some lines about it. This function takes
the fitted parameters of a GEV distribution and calculates
its density at specified sites z.
}
\author{
Philipp Mueller
}

