% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{gpd.density}
\alias{gpd.density}
\title{Port of the ismev::gpd.dens function}
\usage{
gpd.density(parameters, threshold, z)
}
\arguments{
\item{parameters}{Fitted GP parameters c( scale, shape )}

\item{threshold}{Threshold used to fit the GP distribution.}

\item{z}{Numerical vector of sites where to evaluate the
density of the GP distribution.}
}
\value{
Numerical vector of same length as z.
}
\details{
It might not to be called by the user, but I can
certainly write some lines about it. This function takes
the fitted parameters of a GP distribution and calculates
its density at specified sites z.
}
\author{
Philipp Mueller
}
\seealso{
Other extremes: \code{\link{block}},
  \code{\link{decluster}}, \code{\link{extremal.index}},
  \code{\link{gev.density}}, \code{\link{qevd}},
  \code{\link{return.level}}, \code{\link{revd}},
  \code{\link{rlevd}}, \code{\link{threshold}}
}

