% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_leaflet.R
\name{leafletClimex}
\alias{leafletClimex}
\title{Leaflet interfaces to select stations or visualize spacial
information.}
\usage{
leafletClimex(input, output, session, reactive.chosen, buttonMinMax,
  radioEvdStatistics, sliderYears, data.blocking, evd.fitting, sliderThreshold,
  fit.interactive, cleaning.interactive, deseasonalize.interactive,
  blocking.interactive)
}
\arguments{
\item{input}{Namespace input. For more details check out
\link{ \url{ http://shiny.rstudio.com/articles/modules.html } }}

\item{output}{Namespace output.}

\item{session}{Namespace session.}

\item{reactive.chosen}{Reactive value containing a list of the list
of all provided stations and a data.frame containing the meta data.}

\item{buttonMinMax}{Character (radio) input determining whether
the GEV/GP distribution shall be fitted to the smallest or biggest
vales. Choices: c( "Max", "Min ), default = "Max".}

\item{radioEvdStatistics}{Character (radio) input determining whether
the GEV or GP distribution shall be fitted to the data. Choices:
c( "GEV", "GP" ), default = "GEV".}

\item{sliderYears}{Numerical (slider) input to determine the minimal
length (in years) of the time series to be displayed. Minimal value
is 0 and maximal is 155 (longest one in the DWD database), the
default value is 65 and the step width is 1.}

\item{data.blocking}{Reactive value returning a list containing three
elements: 1. the blocked time series, 2. the deseasonalized time
series, and 3. the pure time series.}

\item{evd.fitting}{Reactive value containing the results of the fit
(\code{\link{fit.gev}} or \code{\link{fit.gpd}} depending on
radioEvdStatistic) to the blocked time series in
data.blocking()[[ 1 ]].}

\item{sliderThreshold}{Numerical (slider) input determining the
threshold used within the GP fit and the extraction of the extreme
events. Boundaries: minimal and maximal value of the deseasonalized
time series (rounded). Default: 0.8* the upper end point.}

\item{fit.interactive}{Function used to perform the actual GEV/GP
fit.}

\item{cleaning.interactive}{Function used to remove incomplete years
from blocked time series or to remove clusters from data above a
certain threshold.}

\item{deseasonalize.interactive}{Function used to remove seasonality
from a given time series.}

\item{blocking.interactive}{Function used to split a time series into
blocks of equal lengths and to just extract the maximal values from
then or to extract all data points above a certain threshold value.
Which option is chosen depends of the radioEvdStatistic}
}
\value{
Reactive value holding the selected station.
}
\details{
This module provides an interactive map to display the
locations of the individual stations. The user can choose
individual stations by clicking at them. In addition a dialog
will pop up telling the stations name, the length of the
time series and the 20, 50 and 100 year return level calculated
with the setting in the basic map without station positions        
select only those stations in Germany with a certain minimum
number of years.
}
\author{
Philipp Mueller
}

