% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_leaflet.R
\name{leafletClimexUI}
\alias{leafletClimexUI}
\title{Leaflet interfaces to select stations or visualize spacial
information.}
\usage{
leafletClimexUI(id)
}
\arguments{
\item{id}{Namespace prefix}
}
\value{
tagList
}
\details{
This function provides the user interface to
\code{\link{leafletClimex}}. It consists of a leafletOutput and two
absolutePanels containing the return level and time series length
sliders as well as the table displaying all the stations information.
}
\author{
Philipp Mueller
}
\seealso{
Other leaflet: \code{\link{data.chosen}},
  \code{\link{leafletClimex}}
}

