% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti.R
\name{likelihood.gradient.augmented}
\alias{likelihood.gradient.augmented}
\title{Calculated the gradient of the augmented negative log
  likelihood of the GEV or GP function.}
\usage{
likelihood.gradient.augmented(parameters, x.in, model = c("gev", "gpd"),
  lagrangian.multiplier = rep(0, length(x.in) + 2),
  penalty.parameter = 1000)
}
\arguments{
\item{parameters}{Numerical vector containing the location, scale,
and shape parameters for the GEV or the scale and shape
parameters for the GP. If NULL,
\code{\link{likelihood.initials}} is used to determine
them. Default = NULL}

\item{x.in}{Time series of class \pkg{xts}.}

\item{model}{String determining whether to calculate the initial
parameters of the GEV ("gev") or GP ("gpd") function. Default =
"gev".}

\item{lagrangian.multiplier}{Lagrangian multipliers used to weight
the linear contribution of the constraints. In most cases all of
them are zero, since optimization of the GEV/GP likelihood
usually doesn't take place inside a region of constraint
violations. When supplying this parameter it has to have the
same length as present number of constraints: number of points
in x.in + 2, with the last two constraints handling the lower
bound of the scale and the shape parameter. Default = 0 for all
constraints.}

\item{penalty.parameter}{Penalty parameter used to weight the
quadratic contribution of the constraints. In the end of a
typical constrained GEV or GP optimization this parameter is
1000. Default = 1000.}
}
\value{
Numerical value of the gradient of the augmented negative
  log likelihood.
}
\description{
This function uses the
  \code{\link{likelihood.gradient}} function and adds the linear
  constraints used in \code{\link{fit.gev}} and
  \code{\link{fit.gpd}} to produce the augmented Lagrangian
  version of the GEV or GP negative log-likelihood function.
}
\details{
A convenience function not used by the fitting routines.

  It is only meant to work with constant parameters and no
  covariates.
}
\seealso{
Other optimization: \code{\link{fit.gev.list}},
  \code{\link{fit.gev.xts}}, \code{\link{fit.gev}},
  \code{\link{fit.gpd.list}}, \code{\link{fit.gpd.xts}},
  \code{\link{fit.gpd}},
  \code{\link{likelihood.augmented}},
  \code{\link{likelihood.gradient}},
  \code{\link{likelihood.initials}},
  \code{\link{likelihood}}
}
\author{
Philipp Mueller
}
