% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti.R
\name{likelihood.gradient}
\alias{likelihood.gradient}
\title{Calculates the gradient of the negative log likelihood of the GEV or GPD function.}
\usage{
likelihood.gradient(parameters, x.in, model = c("gev", "gpd"))
}
\arguments{
\item{parameters}{Vector containing the location, scale and shape parameter for the GEV model or the scale and shape parameter for the GPD one.}

\item{x.in}{Time series or numerical vector containing the extreme events.}

\item{model}{Determining whether to calculate the initial parameters of the GEV or GPD function. Default = "gev".}
}
\value{
Numerical vector containing the derivative of the negative log likelihood in (location, scale, shape for GEV) or (scale, shape for GPD) direction.
}
\details{

}
\author{
Philipp Mueller
}
\seealso{
Other optimization: \code{\link{fit.gev}},
  \code{\link{fit.gpd}}, \code{\link{likelihood.initials}},
  \code{\link{likelihood}}
}

