% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{likelihood.gui}
\alias{likelihood.gui}
\title{likelihood.gui}
\usage{
likelihood.gui(time.series, starting.points = NULL, scale.lim = NULL,
  shape.lim = NULL, location.offset = NULL,
  optimization.function = nmk.modified, optimization.steps = c(0.1, 0.5),
  likelihood.minimum = NULL)
}
\arguments{
\item{time.series}{Vector of block maxima.}

\item{starting.points}{Data.frame of the starting points where each one is contained in the single row and the columns are spanned by location, scale and shape. If NULL 7 random points are drawn randomly in the region spanned by scale.lim and shape.lim with the parameter value of the location plane. Default = NULL}

\item{scale.lim}{Region of the scale parameter for which the likelihood function is going to be evaluated and plotted as a density plot and a contour plot. If NULL the MLE estimate using the time series will be calculated an the limits will be set to +/- 5 this value. Default = NULL.}

\item{shape.lim}{Region of the shape parameter for which the likelihood function is going to be evaluated and plotted as a density plot and a contour plot. If NULL the MLE estimate using the time series will be calculated an the limits will be set to +/- 5 this value. Default = NULL.}

\item{location.offset}{Use only if another location plane should be drawn. Default = NULL}

\item{optimization.function}{Function use for preforming the GEV fit. This must be a function providing the list element "x.update" containing the parameters evaluated at each step of the optimization procedure and a list element "par" containing a vector of the GEV parameter estimates. For now only the function nmk.modified is available. Default = nmk.modified.}

\item{optimization.steps}{Vector containing two numbers from 0 to 1 specifying the start and the end point of the optimization. Since the number of steps is unknown beforehand it will be chosen relativley to the total number of steps. Default = c( 0, 1 ).}

\item{likelihood.minimum}{The true minimum of the likelihood function. This should be supplied if gev.fit( x ) doesn't find the optimum in the first run. Default = NULL.}
}
\value{
Opens a HTML widget showing the animation of the optimization routine.
}
\description{
Displays a contour plot of the likelihood function of a time series and the optimization routes for a bunch of provided initial points. This debugging tool is supposed to shed some light on failing optimization runs.
}
\details{
A 2D plot is done in the scale and in the shape plane of the negative log-likelihood. The GEV fit is done by the chosen optimization routine and the plane of the optimal location parameter is chosen for the plot. If any other should be used instead use 'location.offset'. But the trajectories will move out of this plane and so the contour plot is not valid anymore. This function will produce a folder "images" containing the individual images of the optimization. Also a html file containing the animation will be produced. The random initial points are plotted colourful and the heuristic estimate of the starting point will always be plotted in black as well. If the number of starting.points is reduced to 1 another mode of visualization will be triggered. Then the likelihood plane will be updated in every step. So one is able to determine the influence of the location parameter on the fit as well.
}
\author{
Philipp Mueller
}
\seealso{
Other shiny: \code{\link{climex}}
}

