% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti.R
\name{likelihood.initials}
\alias{likelihood.initials}
\title{Estimates the initial GEV parameters of a time series.}
\usage{
likelihood.initials(x, type = c("best", "mom", "lmom"), modified = TRUE,
  restrict = FALSE, restrict.thresholds = c(1000, 500, 3))
}
\arguments{
\item{x}{Time series/numeric.}

\item{type}{Which method should be used to calculate the initial parameters. "best" combines all methods and return the result with the least likelihood. "mom" - method of moments returns an approximation according to the first two moments of the Gumbel distribution. "lmom" - returns an estimate according to the Lmoments method. Default = "best"}

\item{modified}{If TRUE my changes in the heuristics will be applied. Like using the skewness etc. Default = FALSE}

\item{restrict}{If TRUE the estimates will be minimum of the two different approximations and restrict.thresholds. This should prevent artifacts due to diverging means and variances. Default = FALSE}

\item{restrict.thresholds}{Maximal value for each of the estimates. Default = c( 50, 50, 5 )}
}
\value{
Numerical vector containing the c( location, scale, shape ) estimate.
}
\details{
Two main methods are used for the estimation: the L-moments method of Hosking & Wallis implemented in the extRemes::initializer.lmoments() function and an estimation using the first two moments of the Gumbel distribution. For the later one a modification was added: By looking at skewness of the distribution and with respect to some heuristic thresholds a shape parameter between -.4 and .2 is assigned. Warning: both methods do not work for samples with diverging (or pretty big) mean or variance. For this reason the restrict argument is included. If the estimates are bigger than the corresponding restrict.thresholds, they will be replaced by this specific value.
}
\author{
Philipp Mueller
}
\seealso{
Other optimization: \code{\link{fit.gev}},
  \code{\link{fit.gpd}}, \code{\link{likelihood.gradient}},
  \code{\link{likelihood}}
}

