% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_animation.R
\name{likelihoodAnimationUI}
\alias{likelihoodAnimationUI}
\title{User interface controlling the optimization's animation}
\usage{
likelihoodAnimationUI(id)
}
\arguments{
\item{id}{Namespace prefix}
}
\value{
tagList
}
\details{
This function generates two boxes containing the
table of all starting points, the animations/pictures, as well
as all the sliders, numerical inputs etc. to control the
animation.
}
\author{
Philipp Mueller
}
\seealso{
Other animation: \code{\link{animation.wrapper}},
  \code{\link{likelihoodAnimation}},
  \code{\link{plot.animation}}
}

