% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules2.R
\name{plot.animation}
\alias{plot.animation}
\title{Displays the contour plots of the GEV likelihood function of a time series and the optimization routes for a bunch of provided initial points.}
\usage{
\method{plot}{animation}(time.series, starting.points, location.lim = NULL,
  scale.lim = NULL, shape.lim = NULL,
  optimization.function = climex:::nmk.modified, optimization.steps = c(0.1,
  0.5), height = 300, width = 300, colors = list(plane.low = "skyblue1",
  plane.high = "#191970", plane.contour = "white", path.low = "yellow",
  path.high = "darkred", path.true = "black"), folder.name = "images")
}
\arguments{
\item{time.series}{Vector of block maxima.}

\item{starting.points}{Data.frame of the starting points where each one is contained in the single row and the columns are spanned by location, scale and shape.}

\item{location.lim}{Region of the location parameter for which the likelihood function is going to be evaluated and plotted as a density plot and a contour plot.}

\item{scale.lim}{Region of the scale parameter for which the likelihood function is going to be evaluated and plotted as a density plot and a contour plot.}

\item{shape.lim}{Region of the shape parameter for which the likelihood function is going to be evaluated and plotted as a density plot and a contour plot.}

\item{optimization.function}{Function use for preforming the GEV fit. This must be a function providing the list element "x.update" containing the parameters evaluated at each step of the optimization procedure and a list element "par" containing a vector of the GEV parameter estimates. For now only the function nmk.modified is available. Default = nmk.modified.}

\item{optimization.steps}{Vector containing two numbers from 0 to 1 specifying the start and the end point of the optimization. Since the number of steps is unknown beforehand it will be chosen relativley to the total number of steps. Default = c( 0, 1 ).}

\item{height}{of both the images and the form with the playback options in pixel.}

\item{width}{of both the images and the form with the playback options in pixel.}

\item{colors}{List of colors used to generate the plots.}

\item{folder.name}{where the generated pictures should be saved.}
}
\value{
Opens a HTML widget showing the animation of the optimization routine.
}
\details{
Three orthogonal 2D plots are done for the negative log-likelihood of the GEV function intersecting in the actual result of the default optimization. Caution: the trajectories will move out of the planes and so the precise position of the trajectory might be misleading. But the overall goal is to check for local minima. A bunch of images will be generated in the provided folder.
}
\author{
Philipp Mueller
}
\seealso{
Other shiny: \code{\link{animation.wrapper}},
  \code{\link{climex}}
}

