% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.climex.fit.gev}
\alias{print.climex.fit.gev}
\title{Print details}
\usage{
\method{print}{climex.fit.gev}(x, ...)
}
\arguments{
\item{x}{Object of class \emph{climex.fit.gev}.}

\item{...}{Additional parameters. They won't be handled in the
function. This argument is only present to ensure S3 generic
consistency with respect to the \code{\link[base]{print}}
function.}
}
\value{
\code{invisible()}
}
\description{
Summary of the GEV fit results
}
\author{
Philipp Mueller
}
