% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.climex.fit.gpd}
\alias{print.climex.fit.gpd}
\title{Print results}
\usage{
\method{print}{climex.fit.gpd}(x, ...)
}
\arguments{
\item{x}{Object of class \emph{climex.fit.gpd}.}

\item{...}{Additional parameters. They won't be handled in the
function. This argument is only present to ensure S3 generic
consistency with respect to the \code{\link[base]{print}}
function.}
}
\value{
\code{invisible()}
}
\description{
Summary of the GP fit results
}
\author{
Philipp Mueller
}
