% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hacks.R
\name{rcgmin}
\alias{rcgmin}
\title{rcgmin}
\usage{
rcgmin(par = likelihood.initials(x), fn = likelihood,
  gr = likelihood.gradient, x, control = list(), ...)
}
\arguments{
\item{par}{Initial parameters.}

\item{fn}{Function to be optimized (neglog likelihood of the GEV).}

\item{gr}{Gradient of the function to be optimized.}

\item{control}{Additional settings for the CG algorithms}

\item{...}{Additional arguments for fn}
}
\value{
List. Containing e.g. the fitted parameters in the element 'par'.
}
\description{
R version of conjugated gradient algorithm
}
\details{
Modified for my own needs (e.g. x as an input)
}
\author{
Philipp Mueller
}

