% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{remove.incomplete.years}
\alias{remove.incomplete.years}
\title{Remove incomplete years}
\usage{
remove.incomplete.years(x, time.unit = 1)
}
\arguments{
\item{x}{Either a \pkg{xts} class object or a list of those.}

\item{time.unit}{Minimal differences found when applying the
\code{\link[base]{diff}} function to \strong{x}. Default = 1.}
}
\value{
Object of the same class as the input \strong{x}.
}
\description{
Removes all years, which contain either a NA or are
  incomplete (derived from the amount of time stamps within a
  year).
}
\details{
Since incomplete years are detected via the
  difference in the time stamps of neighbouring points, the user
  has to provide the basic time unit. (Numerical value of the
  result of the \code{\link[base]{diff}} function applied to two
  consecutive points.) For daily data the value is 1.
}
\seealso{
\code{\link{remove.seasonality}}

Other ts: \code{\link{aic.climex.fit.gev}},
  \code{\link{aic.list}}, \code{\link{aic}},
  \code{\link{anomalies.list}},
  \code{\link{anomalies.xts}}, \code{\link{anomalies}},
  \code{\link{bic.climex.fit.gev}}, \code{\link{bic.list}},
  \code{\link{bic}}, \code{\link{check.completeness}},
  \code{\link{remove.incomplete.years.list}},
  \code{\link{remove.incomplete.years.xts}},
  \code{\link{remove.seasonality.list}},
  \code{\link{remove.seasonality.xts}},
  \code{\link{remove.seasonality}}
}
\author{
Philipp Mueller
}
