% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{remove.seasonality.xts}
\alias{remove.seasonality.xts}
\title{Remove seasonality}
\usage{
\method{remove.seasonality}{xts}(x)
}
\arguments{
\item{x}{An object of class \pkg{xts}.}
}
\value{
Some class as input.
}
\description{
Calculates the seasonal component of a time series
  and subtracts it from the original.
}
\details{
The function \code{\link[stats]{stl}} with the argument
  \code{s.window = 12} and a 
  conversion of the input into a \emph{ts} class object of daily
  data is used to calculate the seasonal component. This should be
  replaced by a more sophisticated solution as soon I digged
  deeper into the field of deseasonalization.
  \code{\link{remove.incomplete.years}} is used
  to remove incomplete years from the data set. This ensures a
  better calculation of the seasonal component but also requires
  to forecast it to the length of the original data set and align
  it at the right place for subtraction.

  This function can also be applied to a list of \pkg{xts} class
  objects.
}
\seealso{
Other ts: \code{\link{aic.climex.fit.gev}},
  \code{\link{aic.list}}, \code{\link{aic}},
  \code{\link{anomalies.list}},
  \code{\link{anomalies.xts}}, \code{\link{anomalies}},
  \code{\link{bic.climex.fit.gev}}, \code{\link{bic.list}},
  \code{\link{bic}}, \code{\link{check.completeness}},
  \code{\link{remove.incomplete.years.list}},
  \code{\link{remove.incomplete.years.xts}},
  \code{\link{remove.incomplete.years}},
  \code{\link{remove.seasonality.list}},
  \code{\link{remove.seasonality}}
}
\author{
Philipp Mueller
}
