% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{return.level.numeric}
\alias{return.level.numeric}
\title{Calculation of the return levels.}
\usage{
\method{return.level}{numeric}(x, return.period = 100,
  error.estimation = c("none", "MC", "MLE", "bootstrap"),
  model = c("gev", "gpd"), monte.carlo.sample.size = 100,
  bootstrap.sample.size = 100, threshold = NULL, total.length = NULL,
  thresholded.time.series = NULL, extreme.type = c("max", "min"),
  silent = FALSE, mc.cores = NULL, ...)
}
\arguments{
\item{x}{Of class \emph{numeric}.}

\item{return.period}{Numeric vector of the return periods in
years. Default = 100.}

\item{error.estimation}{The estimation of the fitting error is
only for input argument of the class \emph{climex.fit.gev} or
\emph{climex.fit.gpd}. Else the correct scaling of the can not
be assured.}

\item{model}{String determining whether to calculate the initial
parameters of the GEV ("gev") or GP ("gpd") function. Default =
"gev"}

\item{monte.carlo.sample.size}{Number of samples used to obtain
the Monte Carlo estimate of the standard error of the fitting.
Default = 100.}

\item{bootstrap.sample.size}{Number of samples with replacements
to drawn from the original series \strong{x} in order to
determine the standard errors for the GEV/GP parameters and
return levels. Default = 100.}

\item{threshold}{Optional threshold for the GP model. If present
it will be added to the return level to produce a value which
fits to underlying time series. Default = NULL.}

\item{total.length}{Total number of observations in the time
series the exceedance were obtained from (before! applying the
threshold). This argument is needed to calculate the standard
error of the return level via the delta method of the MLE in the
GP model. Default = NULL.}

\item{thresholded.time.series}{Time series used with
\code{\link{fit.gpd}} on which already a threshold (the one
supplied here as well) was applied. Necessary to transform the
return level for numerical input and the GP model from m-th
observation return level to annual return level. If omitted the
return level will be per observation. Default = NULL.}

\item{extreme.type}{String indicating whether to calculate the
quantiles at the right ("max") or left ("min") side of the PDF
of the series. Default = "max".}

\item{silent}{Throws an warning whenever the "gpd" model is used
and the \strong{thresholded.time.series} is not supplied. Since
this can be annoying one can also disable it. Default = FALSE.}

\item{mc.cores}{A numerical input specifying the number of cores
to use for the multi core application of the function (see
\code{\link[parallel]{detectCores}}). This functionality is only
available if the input is a list of different objects. If NULL,
the function will be calculated classically with only one core.
Default = NULL.}
}
\value{
A list containing the estimates "return.level" and their
  standard errors "error" if the input was a fitting object or a
  numerical input.
}
\description{
Calculate arbitrary return levels and their error
  estimates for generalized extreme value (GEV) and generalized
  Pareto (GP) distributions.
}
\details{
Uses the \code{\link{rlevd}} function at its core (a port
  from the \pkg{extRemes} package) but is also capable of
  calculating numerous return levels at once, and also calculates
  the errors of the return levels. For the errors the ML fit is
  using the option \code{hessian=TRUE} (if not done already) or
  uses a Monte Carlo based approach. If no fitting object is
  provided, no errors will be calculated.
}
\examples{
fit.results <- fit.gev( block( anomalies( temp.potsdam ) ) )
return.level( fit.results, return.period = c( 10, 50, 100 ),
              error.estimation = "MLE" )
}
\seealso{
Other extremes: \code{\link{block.list}},
  \code{\link{block.xts}}, \code{\link{block}},
  \code{\link{decluster.list}},
  \code{\link{decluster.xts}}, \code{\link{decluster}},
  \code{\link{extremal.index}}, \code{\link{gev.density}},
  \code{\link{gpd.density}}, \code{\link{qevd}},
  \code{\link{return.level.climex.fit.gev}},
  \code{\link{return.level.climex.fit.gpd}},
  \code{\link{return.level.list}},
  \code{\link{return.level}}, \code{\link{revd}},
  \code{\link{rlevd}}, \code{\link{threshold.list}},
  \code{\link{threshold.xts}}, \code{\link{threshold}}
}
\concept{extremes}
