% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{revd}
\alias{revd}
\title{Drawing random numbers from the GEV or GP distribution}
\usage{
revd(n, location = NULL, scale = NULL, shape = NULL, threshold = NULL,
  model = c("gev", "gpd"), silent = FALSE)
}
\arguments{
\item{n}{Number of samples to draw}

\item{location}{Of the GEV distribution. Default = NULL.}

\item{scale}{Of the GEV/GP distribution. Default = NULL.}

\item{shape}{Of the GEV/GP distribution. Default = NULL.}

\item{threshold}{Used in the GP distribution. This parameter is
optional but should be provided in order to create a representation
of the fitted data exceedance. Default = NULL.}

\item{model}{Determines if to use the GEV or GP distribution.
Default = "gev".}

\item{silent}{Whether to display warnings or not. Default = FALSE.}
}
\value{
Numerical vector of length n drawn from the corresponding
distribution.
}
\details{
This function was originally part of the extRemes package.
But since one had to provide the threshold I couldn't use it insight
the fit.gpd function. In contrast to the original implementation this
function only features constant location, scale and shape parameters.
If you want to do time dependent analysis of extreme events please
refer to the original package.
}
\seealso{
Other extremes: \code{\link{block}},
  \code{\link{decluster}}, \code{\link{extremal.index}},
  \code{\link{gev.density}}, \code{\link{gpd.density}},
  \code{\link{qevd}}, \code{\link{return.level}},
  \code{\link{rlevd}}, \code{\link{threshold}}
}
\author{
Philipp Mueller
}
