% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarCleaning}
\alias{sidebarCleaning}
\title{Toggling the cleaning of the time series}
\usage{
sidebarCleaning(radioEvdStatistics)
}
\arguments{
\item{radioEvdStatistics}{Character (radio) input determining whether
the GEV or GP distribution shall be fitted to the data. Choices:
c( "GEV", "GP" ), default = "GEV".}
}
\value{
renderMenu
}
\details{
A checkbox input. If toggled and the GEV distribution was
chosen via input$radioEvdStatistics, all incomplete years will be
removed from the time series. If on the other hand the GP
distribution was picked, all cluster will be removed and only their
highest point will remain.
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{file.loading}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}}
}
\author{
Philipp Mueller
}
