% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarDataBase}
\alias{sidebarDataBase}
\title{Selecting the database}
\usage{
sidebarDataBase()
}
\value{
renderMenu
}
\details{
For now there are three different data sources you can
choose:the input provided when calling climex::climex() on localhost
or a file loaded via the sidebar, the station data of the German
weather service (DWD) and artificial data sampled from a GEV
distribution. Beware: it's not a true module! Using the namespaces
does not really make sense because it has to be accessed from outside
of this context.
}
\author{
Philipp Mueller
}
\seealso{
Other sidebar: \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}}
}

