% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarDataBaseInput}
\alias{sidebarDataBaseInput}
\title{Selecting the database in the Climex app}
\usage{
sidebarDataBaseInput()
}
\value{
menuItemOutput
}
\description{
Selecting the database.
}
\details{
Provides the \code{\link[shinydashboard]{menuItemOutput}}
  for \code{\link{sidebarDataBase}}. See the later one for
  details.
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{file.loading}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarImprintInput}},
  \code{\link{sidebarImprint}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}},
  \code{\link{sidebarSeriesLengthInput}},
  \code{\link{sidebarSeriesLength}}
}
\author{
Philipp Mueller
}
