% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarDataSource}
\alias{sidebarDataSource}
\title{Sidebar menu selection}
\usage{
sidebarDataSource(selectDataBase, radioEvdStatistics, reactive.chosen,
  selected.station)
}
\arguments{
\item{selectDataBase}{Character (select) input to determine the data
source. In the default installation there are three options:
c( "input", "DWD", "artificial data" ). The first one uses the data
provided as an argument to the call of the \code{\link{climex}}
function. The second one uses the database of the German weather
service (see \code{link{download.data.dwd}}). The third one allows
the user to produce random numbers distributed according to the GEV
or GP distribution. Determined by menuSelectDataBase.
Default = "DWD".}

\item{radioEvdStatistics}{Character (radio) input determining whether
the GEV or GP distribution shall be fitted to the data. Choices:
c( "GEV", "GP" ), default = "GEV".}

\item{reactive.chosen}{Reactive value containing a list of the list
of all provided stations and a data.frame containing the meta data.}

\item{selected.station}{Reactive value provided by the \code{\link{
leafletClimex}} function. It contains the name of the station selected
by the user by clicking on the leaflet map. It's provided as a
character and NULL if no click occurred so far.}
}
\value{
renderMenu
}
\details{
Provides the second selection menu in the sidebar.
If input$selectDataBase, provided by \code{\link{sidebarDataBase}},
was set to "artificial data", the menu will be a numerical input
slider to provide the location parameter for the GEV or scale
parameter for the GP distribution. Else it will be a drop down menu
listing the names of all available stations (see \code{\link{
data.chosen}})
}
\author{
Philipp Mueller
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}}
}

