% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarDataSourceInput}
\alias{sidebarDataSourceInput}
\title{Sidebar menu selection}
\usage{
sidebarDataSourceInput()
}
\value{
menuItemOutput
}
\details{
Provides the shinydashboard::menuItemOutput for \code{\link{
sidebarDataSource}}. See the later one for details
}
\author{
Philipp Mueller
}
\seealso{
Other sidebar: \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}}
}

