% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarDataSourceInput1}
\alias{sidebarDataSourceInput1}
\title{Sidebar menu selection}
\usage{
sidebarDataSourceInput1()
}
\value{
menuItemOutput
}
\details{
Provides the shinydashboard::menuItemOutput for \code{\link{
sidebarDataSource1}}. See the later one for details
}
\author{
Philipp Mueller
}
\seealso{
Other sidebar: \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSource1}}
}

