% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarDataType}
\alias{sidebarDataType}
\title{Sidebar menu selection}
\usage{
sidebarDataType(selectDataBase, radioEvdStatistics)
}
\arguments{
\item{selectDataBase}{Character (select) input to determine the data
source. In the default installation there are three options:
c( "input", "DWD", "artificial data" ). The first one uses the data
provided as an argument to the call of the \code{\link{climex}}
function. The second one uses the database of the German weather
service (see \code{link{download.data.dwd}}). The third one allows
the user to produce random numbers distributed according to the GEV
or GP distribution. Determined by menuSelectDataBase.
Default = "DWD".}

\item{radioEvdStatistics}{Character (radio) input determining whether
the GEV or GP distribution shall be fitted to the data. Choices:
c( "GEV", "GP" ), default = "GEV".}
}
\value{
renderMenu
}
\details{
Provides the third selection menu in the sidebar.
If input$selectDataBase, provided by \code{\link{sidebarDataBase}},
was set to "artificial data", the menu will be a numerical input
slider to provide the scale parameter for the GEV or shape
parameter for the GP distribution. Else it will be a drop down menu
listing the different types of data available at the chosen station.
}
\author{
Philipp Mueller
}
\seealso{
Other sidebar: \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}}
}

