% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarImprint}
\alias{sidebarImprint}
\title{Linking the imprint of the climex application}
\usage{
sidebarImprint(session)
}
\arguments{
\item{session}{Namespace session. For details check out
\link{ \url{ http://shiny.rstudio.com/articles/modules.html}}}
}
\value{
div
}
\details{
Whenever the web app is run using shiny-server a link to
an arbitrary HTML page (preferably an imprint) will be rendered.
Else just an empty div will be added.
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{file.loading}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarImprintInput}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}},
  \code{\link{sidebarSeriesLengthInput}},
  \code{\link{sidebarSeriesLength}}
}
\author{
Philipp Mueller
}
