% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarLoadingGif}
\alias{sidebarLoadingGif}
\title{Showing a Gif in the sidebar while the app is processing/on hold}
\usage{
sidebarLoadingGif(input, output, session)
}
\arguments{
\item{input}{Namespace input. For more details check out
\link{ \url{ http://shiny.rstudio.com/articles/modules.html } }}

\item{output}{Namespace output.}

\item{session}{Namespace session.}
}
\value{
div
}
\details{
This module will display a loading gif. Most of the code of
this function is already written in a JavaScript script installed
alongside the climex package. It just determines if the app is run on
either a localhost (direct interaction) or on a server waiting for
client interaction. Accordingly it looks up both the JavaScript script
and the corresponding gif file.
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{file.loading}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarImprintInput}},
  \code{\link{sidebarImprint}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}},
  \code{\link{sidebarSeriesLengthInput}},
  \code{\link{sidebarSeriesLength}}
}
\author{
Philipp Mueller
}
