% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarLoadingGifOutput}
\alias{sidebarLoadingGifOutput}
\title{Showing a Gif in the sidebar while the app is processing/on hold}
\usage{
sidebarLoadingGifOutput(id)
}
\arguments{
\item{id}{Namespace prefix}
}
\value{
div
}
\details{
Rendering of \code{link{sidebarLoadingGif}}.
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{file.loading}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarImprintInput}},
  \code{\link{sidebarImprint}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}},
  \code{\link{sidebarSeriesLengthInput}},
  \code{\link{sidebarSeriesLength}}
}
\author{
Philipp Mueller
}
