% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarSeriesLengthInput}
\alias{sidebarSeriesLengthInput}
\title{Series length in the Climex app}
\usage{
sidebarSeriesLengthInput()
}
\value{
menuItemOutput
}
\description{
Slider to determine the length of an artificial time series
}
\details{
Provides the \code{\link[shinydashboard]{menuItemOutput}}
  for \code{\link{sidebarSeriesLength}}. See the later one for
  details.
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{file.loading}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarImprintInput}},
  \code{\link{sidebarImprint}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}},
  \code{\link{sidebarSeriesLength}}
}
\author{
Philipp Mueller
}
