% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.climex.fit.gev}
\alias{summary.climex.fit.gev}
\title{Summarize details}
\usage{
\method{summary}{climex.fit.gev}(object, ...)
}
\arguments{
\item{object}{Object of class \emph{climex.fit.gev}.}

\item{...}{Additional parameters. They won't be handled in the
function. This argument is only present to ensure S3 generic
consistency with respect to the \code{\link[base]{summary}}
function.}
}
\value{
\code{invisible()}
}
\description{
Summary of the GEV fit results
}
\author{
Philipp Mueller
}
