% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\name{temp.potsdam}
\alias{temp.potsdam}
\title{Data set of daily maximum temperatures of the Potsdam station
provided by the DWD. If you intend to use this data please look at
the \href{ftp://ftp-cdc.dwd.de/pub/CDC/Terms_of_use.pdf}{terms of
use}. The data is stored in the \pkg{xts} format.}
\format{Class \pkg{xts}}
\source{
\url{http://www.dwd.de/}
}
\description{
Data set of daily maximum temperatures of the Potsdam station
provided by the DWD. If you intend to use this data please look at
the \href{ftp://ftp-cdc.dwd.de/pub/CDC/Terms_of_use.pdf}{terms of
use}. The data is stored in the \pkg{xts} format.
}
