% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{threshold.list}
\alias{threshold.list}
\title{Apply a threshold to a time series}
\usage{
\method{threshold}{list}(x, threshold, decluster = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{A list of objects of class \pkg{xts}.}

\item{threshold}{Numerical value, which has to be exceeded by the
data.}

\item{decluster}{Logical flag indicating whether or not to
decluster the obtained exceedances over the
\strong{threshold}. Default = TRUE.}

\item{na.rm}{Logical flag indicating whether to remove all NA
values from the time series (removed points in clusters). For
important steps, like calculating the \emph{Lmoments} of a time
series, there must not be any NA left. Default = TRUE.}
}
\value{
Same class as the input
}
\description{
Extracts all data above a certain threshold of a
  \pkg{xts} class object or a list of those.
}
\details{
After extracting all data above the \strong{threshold},
  the \strong{threshold} argument itself is going to be subtracted
  from the data.

  In addition, all exceedance can be declustered. This step is
  highly recommendes, since the extreme value theory is only valid
  for data without correlations (which are present in most
  measured data). Using this method, at least the short-range
  correlations in the data can be taken care of.
}
\seealso{
Other extremes: \code{\link{block.list}},
  \code{\link{block.xts}}, \code{\link{block}},
  \code{\link{decluster.list}},
  \code{\link{decluster.xts}}, \code{\link{decluster}},
  \code{\link{extremal.index}}, \code{\link{gev.density}},
  \code{\link{gpd.density}}, \code{\link{qevd}},
  \code{\link{return.level.climex.fit.gev}},
  \code{\link{return.level.list}},
  \code{\link{return.level}}, \code{\link{revd}},
  \code{\link{rlevd}}, \code{\link{threshold.xts}},
  \code{\link{threshold}}
}
\author{
Philipp Mueller
}
