% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{threshold}
\alias{threshold}
\title{Apply a threshold to data}
\usage{
threshold(x, threshold, decluster = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{Time series or numerical vector.}

\item{threshold}{Value which has to be exceeded.}

\item{decluster}{Flag indicating whether of not to decluster the
obtained exceedance. Default = TRUE.}

\item{na.rm}{Flag indicating whether to remove all NA values from the
time series (removed points in clusters). For important steps like
calculating the Lmoments of the time series there must not be any NA
left. Default = TRUE.}
}
\value{
Declustered time series of the same format as the input.
}
\description{
Extracts all data above a certain threshold
}
\details{
Due to the UNIX principle (make each program do one thing
well) I decided to provide this extra function instead of
incorporating it into the fitting function. After extracting all data
above the threshold it is going to be subtracted from the data. In
addition all exceedance can be declustered. This step is highly
recommended since the extreme value theory is only valid for data
without correlations and short-range correlations (which are present
in most measured data) can be filtered out using this procedure.
}
\seealso{
Other extremes: \code{\link{block}},
  \code{\link{decluster}}, \code{\link{extremal.index}},
  \code{\link{gev.density}}, \code{\link{gpd.density}},
  \code{\link{qevd}}, \code{\link{return.level}},
  \code{\link{revd}}, \code{\link{rlevd}}
}
\author{
Philipp Mueller
}
