% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{threshold.xts}
\alias{threshold.xts}
\title{Apply a threshold to a time series}
\usage{
\method{threshold}{xts}(x, threshold, decluster = TRUE,
  cluster.distance = NULL, extreme.type = c("max", "min"),
  na.rm = TRUE, mc.cores = NULL)
}
\arguments{
\item{x}{An object of class \pkg{xts}.}

\item{threshold}{Numerical value, which has to be exceeded by the
data.}

\item{decluster}{Logical flag indicating whether or not to
decluster the obtained exceedances over the
\strong{threshold}. Default = TRUE.}

\item{cluster.distance}{Numerical value specifying how many points
have to be below the \strong{threshold} for the next point to be
considered the starting point of a new cluster. Only supply a
value when you really know what you are doing! Default = NULL.}

\item{extreme.type}{String specifying whether the events below a
very low ("min") or above a very high ("max") \strong{threshold}
should be extracted. The \strong{threshold} will be subtracted
from the data in both cases. Default: "max".}

\item{na.rm}{Logical flag indicating whether to remove all NA
values from the time series (removed points in clusters). For
important steps, like calculating the \emph{Lmoments} of a time
series, there must not be any NA left. Default = TRUE.}

\item{mc.cores}{A numerical input specifying the number of cores
to use for the multi core application of the function (see
\code{\link[parallel]{detectCores}}). This functionality is only
available if the input is a list of different objects. If NULL,
the function will be calculated classically with only one core.
Default = NULL.}
}
\value{
Same class as the input
}
\description{
Extracts all data above a certain threshold of a
  \pkg{xts} class object or a list of those.
}
\details{
After extracting all data above the \strong{threshold},
  the \strong{threshold} argument itself is going to be subtracted
  from the data.

  In addition, all exceedance can be declustered. This step is
  highly recommendes, since the extreme value theory is only valid
  for data without correlations (which are present in most
  measured data). Using this method, at least the short-range
  correlations in the data can be taken care of.
}
\seealso{
Other extremes: \code{\link{block.list}},
  \code{\link{block.xts}}, \code{\link{block}},
  \code{\link{decluster.list}},
  \code{\link{decluster.xts}}, \code{\link{decluster}},
  \code{\link{extremal.index}}, \code{\link{gev.density}},
  \code{\link{gpd.density}}, \code{\link{qevd}},
  \code{\link{return.level.climex.fit.gev}},
  \code{\link{return.level.climex.fit.gpd}},
  \code{\link{return.level.list}},
  \code{\link{return.level.numeric}},
  \code{\link{return.level}}, \code{\link{revd}},
  \code{\link{rlevd}}, \code{\link{threshold.list}},
  \code{\link{threshold}},
  \code{\link{upper.limit.climex.fit.gev}},
  \code{\link{upper.limit.climex.fit.gpd}},
  \code{\link{upper.limit.list}},
  \code{\link{upper.limit.numeric}},
  \code{\link{upper.limit}}
}
\author{
Philipp Mueller
}
\concept{extremes}
