% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts.swap}
\alias{ts.swap}
\title{ts.swap}
\usage{
ts.swap(x, number.of.swaps = 1)
}
\arguments{
\item{x}{Input time series. Preferrably of the xts class.}

\item{number.of.swaps}{How many swaps should be performed on the time series. Default = 1;}
}
\value{
Swapped time series of the same class as x.
}
\description{
Swaps one or more entries in the time series.
}
\details{
The swaps is done at random and its not recorded which elements are already swapped.
}
\author{
Philipp Mueller
}
\seealso{
Other ts: \code{\link{aic}}, \code{\link{anomalies}},
  \code{\link{bic}}, \code{\link{remove.incomplete.years}},
  \code{\link{remove.seasonality}}
}

