% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ttplot}
\alias{ttplot}
\title{Plotting a xts time series with ggplot2 in a convenient
  format.}
\usage{
ttplot(data.input, main = "Time series", ylab = NULL)
}
\arguments{
\item{data.input}{Time series which should be visualized.}

\item{main}{Title of the plot to generate. Default: "Time series"}

\item{ylab}{Label of the y-axis of the plot to generate. Default =
NULL}

\item{x.df}{}

\item{...}{Additional parameters for the multiplot function.}
}
\value{
Nothing. The multiplot function is run in the last step.
}
\details{
Plots all objects of class xts. With the main argument a
  title can be provided.
}
\seealso{
Other plot: \code{\link{multiplot}}
}
\author{
Philipp Mueller
}
