% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{bic}
\alias{bic}
\title{Bayesian information criterion}
\usage{
bic(x)
}
\arguments{
\item{x}{Either an object returned by the \code{\link{fit.gev}} or
\code{\link{fit.gpd}} function or a list of such elements.}
}
\value{
Numeric value, if the input is a single fitting object, or
  a numerical vector, if \strong{x} is a list of such objects.
}
\description{
Calculates the Bayesian information criterion of a
  \emph{climex.fit.gev} or \emph{climex.fit.gpd} class object or a
  list of those.
}
\details{
The \emph{climex.fit.gev} object is of identical
  structure as the output of the \code{\link[stats]{optim}}
  function.
}
\seealso{
\code{\link{aic}}, \code{\link{fit.gev}},
  \code{\link{fit.gpd}}

Other ts: \code{\link{aic.climex.fit.gev}},
  \code{\link{aic.list}}, \code{\link{aic}},
  \code{\link{anomalies.list}},
  \code{\link{anomalies.xts}}, \code{\link{anomalies}},
  \code{\link{bic.climex.fit.gev}}, \code{\link{bic.list}},
  \code{\link{check.completeness}},
  \code{\link{remove.incomplete.years.list}},
  \code{\link{remove.incomplete.years.xts}},
  \code{\link{remove.incomplete.years}},
  \code{\link{remove.seasonality.list}},
  \code{\link{remove.seasonality.xts}},
  \code{\link{remove.seasonality}}
}
\author{
Philipp Mueller
}
