% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{bic}
\alias{bic}
\title{Calculates the Bayesian information criterion of a climex.fit.gev or climex.fit.gpd object.}
\usage{
bic(x)
}
\arguments{
\item{x}{Object of class climex.fit.gev or climex.fit.gpd. When ensuring the presence of all additional list elements an output of optim() can also be used.}
}
\value{
Numeric value
}
\details{
The climex.fit.gev object is of identical structure as the output of the optim() function.
}
\author{
Philipp Mueller
}
\seealso{
\code{\link{aic}}

Other ts: \code{\link{aic}}, \code{\link{anomalies}},
  \code{\link{remove.incomplete.years}},
  \code{\link{remove.seasonality}}
}

