% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{bic}
\alias{bic}
\title{Calculates the Bayesian information criterion of a gev.fit or climex.gev.fit object.}
\usage{
bic(x, time.series = NULL)
}
\arguments{
\item{x}{Optim()like fit of a time series.}

\item{time.series}{Original time series.}
}
\value{
Numeric value
}
\details{
The climex.gev.fit object is of identical structure as the output of the optim() function. In contrast to the aic the time series itself has to provided too.
}
\author{
Philipp Mueller
}
\seealso{
\code{\link{aic}}

Other ts: \code{\link{aic}}, \code{\link{anomalies}},
  \code{\link{remove.incomplete.years}},
  \code{\link{remove.seasonality}}
}

