% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{block}
\alias{block}
\title{Blocking data}
\usage{
block(x, block.number = round(length(x)/50), block.length = NULL,
  block.mode = c("max", "min"), separation.mode = c("years", "none"))
}
\arguments{
\item{x}{Either an object of class \pkg{xts} or a list of those.}

\item{block.number}{Specifies the number of blocks the input data is
going to be separated in.}

\item{block.length}{Length of the blocks. For the sake of simplicity
the last block is not forced to match the length of the other
plots.}

\item{block.mode}{This parameter determines if the maximum "max" or
the minimum "min" of a block shall be returned. Default: "max".}

\item{separation.mode}{"years" is used to split the data according to
its date values instead. If \strong{block.length} or
\strong{block.number} is specified this argument will not be
considered and set to "none". Default = "years".}
}
\value{
Same class as the input.
}
\description{
Separates the input into blocks of equal size and
  returns the maximum or minimum of the block as result.
}
\details{
If \strong{separation.mode} is set to "years" the
  data is separated according to it's time stamps. If not, the size
  of a block is determined by the \strong{block.length} argument
  or calculated via the \strong{block.number} argument if the
  previous one is set to NULL. For calculating the mean of
  the blocks, please have a look at the \code{\link[stats]{ave}}
  function.

  This function can be applied to lists of \pkg{xts} class elements
  too.
}
\examples{
  block( temp.potsdam )
}
\seealso{
Other extremes: \code{\link{block.list}},
  \code{\link{block.xts}}, \code{\link{decluster.list}},
  \code{\link{decluster.xts}}, \code{\link{decluster}},
  \code{\link{extremal.index}}, \code{\link{gev.density}},
  \code{\link{gpd.density}}, \code{\link{qevd}},
  \code{\link{return.level.climex.fit.gev}},
  \code{\link{return.level.list}},
  \code{\link{return.level}}, \code{\link{revd}},
  \code{\link{rlevd}}, \code{\link{threshold.list}},
  \code{\link{threshold.xts}}, \code{\link{threshold}}

Other extremes: \code{\link{block.list}},
  \code{\link{block.xts}}, \code{\link{decluster.list}},
  \code{\link{decluster.xts}}, \code{\link{decluster}},
  \code{\link{extremal.index}}, \code{\link{gev.density}},
  \code{\link{gpd.density}}, \code{\link{qevd}},
  \code{\link{return.level.climex.fit.gev}},
  \code{\link{return.level.list}},
  \code{\link{return.level}}, \code{\link{revd}},
  \code{\link{rlevd}}, \code{\link{threshold.list}},
  \code{\link{threshold.xts}}, \code{\link{threshold}}
}
\author{
Philipp Mueller
}
