% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{check.completeness}
\alias{check.completeness}
\title{Checking for N complete years of data}
\usage{
check.completeness(x, number.of.years = 30, mc.cores = NULL)
}
\arguments{
\item{x}{Either a time series of class \pkg{xts} or a list of
them.}

\item{number.of.years}{The minimum number of complete years in the
series. Default = 30.}

\item{mc.cores}{A numerical input specifying the number of cores
to use for the multi core application of the function (see
\code{\link[parallel]{detectCores}}). This functionality is only
available if the input is a list of different objects. If NULL,
the function will be calculated classically with only one core.
Default = NULL.}
}
\value{
If the input is of type \pkg{xts}, the function returns
  TRUE or FALSE. If the input, on the other hand, is a list of
  objects of type \pkg{xts}, it returns a trimmed list
  containing only those elements, which indeed have more than N
  years of complete data.
}
\description{
This function tests whether the series do has at
  least \strong{number.of.years} complete years of data.
}
\details{
It uses the \code{\link{remove.incomplete.years}}
  to check for the completeness of the individual years.
}
\seealso{
Other ts: \code{\link{aic.climex.fit.gev}},
  \code{\link{aic.climex.fit.gpd}}, \code{\link{aic.list}},
  \code{\link{aic}}, \code{\link{anomalies.list}},
  \code{\link{anomalies.xts}}, \code{\link{anomalies}},
  \code{\link{bic.climex.fit.gev}},
  \code{\link{bic.climex.fit.gpd}}, \code{\link{bic.list}},
  \code{\link{bic}},
  \code{\link{remove.incomplete.years.list}},
  \code{\link{remove.incomplete.years.xts}},
  \code{\link{remove.incomplete.years}},
  \code{\link{remove.seasonality.list}},
  \code{\link{remove.seasonality.xts}},
  \code{\link{remove.seasonality}}
}
\author{
Philipp Mueller
}
