% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_base.R
\name{climex}
\alias{climex}
\title{Shiny app combining most of the tools used in extreme value
analysis using GEV fitting via maximum likelihood.}
\usage{
climex(x.input = NULL)
}
\arguments{
\item{x.input}{Time series of one of the formats mentioned in the
details section.}
}
\value{
starts a shiny app.
}
\details{
It possible to provide a time series of type \pkg{xts} to
a list of elements of this type. This app need the its own css file.
This function will be exclusively called when the climex app is run
on localhost. In order to assure its correct behavior the necessary
file/folder structure in the CLIMEX.PATH (global variable which has
to be set beforehand; see vignette for details) will be check and
if necesary generated. The input time series is accepted in three
different formats
\enumerate{
 \item As a single time series in the formatted using the \pkg{xts}
package 
 \item As a named list of various xts time series 
 \item As a list containing two elements;  a list of xts time
series and a data.frame with the columns \emph{longitude},
\emph{latitude}, \emph{altitude}, \emph{name} corresponding to the
stations geo-coordinates, height and name 
}
}
\author{
Philipp Mueller
}
\seealso{
Other shiny: \code{\link{animation.wrapper}},
  \code{\link{plot.animation}}
}

