% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_base.R
\name{climex}
\alias{climex}
\title{Shiny app combining most of the tools used in extreme value analysis using GEV fitting via maximum likelihood.}
\usage{
climex(x.input = NULL)
}
\arguments{
\item{x.input}{Time series of type xts or list of such time series. Default = NULL}
}
\value{
starts a shiny app.
}
\details{
It possible to provide a time series of type xts to a list of elements of this type. This app need the its own css file.
}
\author{
Philipp Mueller
}
\seealso{
Other shiny: \code{\link{animation.wrapper}},
  \code{\link{plot.animation}}
}

