% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_base.R
\name{climex.ui}
\alias{climex.ui}
\title{The Climex app UI}
\usage{
climex.ui(selected = c("Map", "General"))
}
\arguments{
\item{selected}{Choose which tab is supposed to be selected when
starting the app}
}
\value{
HTML code of the interface
}
\description{
The user interface for the \code{\link{climex}}
  function.
}
\details{
Contains all the HTML codes.
}
\seealso{
Other climex: \code{\link{climex.server}},
  \code{\link{climex}}
}
\author{
Philipp Mueller
}
