% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{data.selection}
\alias{data.selection}
\title{Reactive value selecting an individual time series.}
\usage{
data.selection(reactive.chosen, selectDataSource, selectDataBase, sliderYears,
  sliderThreshold, radioEvdStatistics, sliderArtificialDataLocation,
  sliderArtificialDataScale, sliderArtificialDataShape)
}
\arguments{
\item{reactive.chosen}{Reactive value containing a list of the list
of all provided stations and a data.frame containing the meta data.}

\item{selectDataSource}{Menu output in the sidebar. Since this
function should only be triggered when selectDataBase equals "DWD",
this input will be a character string describing the selected
station's name.}

\item{selectDataBase}{Character (select) input to determine the data
source. In the default installation there are three options:
c( "input", "DWD", "artificial data" ). The first one uses the data
provided as an argument to the call of the \code{\link{climex}}
function. The second one uses the database of the German weather
service (see \code{link{download.data.dwd}}). The third one allows
the user to produce random numbers distributed according to the GEV
or GP distribution. Determined by menuSelectDataBase.
Default = "DWD".}

\item{sliderYears}{Numerical (slider) input to determine the minimal
length (in years) of the time series to be displayed. Minimal value
is 0 and maximal is 155 (longest one in the DWD database), the
default value is 65 and the step width is 1.}

\item{sliderThreshold}{Numerical (slider) input determining the
threshold used within the GP fit and the extraction of the extreme
events. Boundaries: minimal and maximal value of the deseasonalized
time series (rounded). Default: 0.8* the upper end point.}

\item{radioEvdStatistics}{Character (radio) input determining whether
the GEV or GP distribution shall be fitted to the data. Choices:
c( "GEV", "GP" ), default = "GEV".}

\item{sliderArtificialDataLocation}{Numerical (slider) input
providing the location parameter to generate an artificial time
series. For input$radioEvdStatistics == "GP" or input$selectDataBase
!= "artificial data" this argument will be NULL.}

\item{sliderArtificialDataScale}{Numerical (slider) input
providing the scale parameter to generate an artificial time
series. For input$selectDataBase
!= "artificial data" this argument will be NULL.}

\item{sliderArtificialDataShape}{Numerical (slider) input
providing the shape parameter to generate an artificial time
series. For input$selectDataBase
!= "artificial data" this argument will be NULL.}
}
\value{
Reactive value containing a 'xts' class time series.
}
\details{
According to the choice in input$selectDataBase an
artificial time series will be sampled or one from a database will be
selected. For the latter one the reactive value
\code{\link{climex:::data.chosen}} will be constulted. The length of
the generated time series is determined by the input$sliderYears
slider in the top right part of the leaflet tab.
}
\author{
Philipp Mueller
}
\seealso{
Other sidebar: \code{\link{file.loading}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}}
}

