% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{decluster}
\alias{decluster}
\title{decluster}
\usage{
decluster(x, threshold)
}
\arguments{
\item{x}{Time series}

\item{threshold}{Has to be set sufficient high to fulfill the asymptotic condition for the GP distribution.}
}
\value{
Declustered values above the provided threshold (xts).
}
\description{
Decluster point over threshold data used for GP fitting.
}
\details{
Inspired by the decluster algorithm used in the extRemes package
}
\author{
Philipp Mueller
}

