% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_fitting.R
\name{fit.interactive}
\alias{fit.interactive}
\title{Function to perform the GEV/GP fit within the Climex app.}
\usage{
fit.interactive(x.kept, radioEvdStatistics, buttonMinMax, sliderThreshold,
  selectDataBase)
}
\arguments{
\item{x.kept}{Time series of class 'xts'. Removing clicked or brushed
values has to be done beforehand.}

\item{radioEvdStatistics}{Character (radio) input determining whether
the GEV or GP distribution shall be fitted to the data. Choices:
c( "GEV", "GP" ), default = "GEV".}

\item{buttonMinMax}{Character (radio) input determining whether
the GEV/GP distribution shall be fitted to the smallest or biggest
vales. Choices: c( "Max", "Min ), default = "Max".}

\item{sliderThreshold}{Numerical (slider) input determining the
threshold used within the GP fit and the extraction of the extreme
events. Boundaries: minimal and maximal value of the deseasonalized
time series (rounded). Default: 0.8* the upper end point.}

\item{selectDataBase}{Character (select) input to determine the data
source. In the default installation there are three options:
c( "Input", "DWD", "Artificial data" ). The first one uses the data
provided as an argument to the call of the \code{\link{climex}}
function. The second one uses the database of the German weather
service (see \code{\link{download.data.dwd}}). The third one allows
the user to produce random numbers distributed according to the GEV
or GP distribution. Determined by menuSelectDataBase.
Default = "DWD".}
}
\value{
Object of class 'climex.fit.gev' or 'climex.fit.gpd', depending
on the choice in input$radioEvdStatistics.
}
\details{
This function does not wait for the initialization of its
slider, checkbox etc. inputs. This way the fit can be performed with its
default settings in the leaflet tab without switching to the General tab
first.
}
\seealso{
Other climex-fitting: \code{\link{color.table}},
  \code{\link{data.fitting}},
  \code{\link{generalFitStatisticsTable}},
  \code{\link{generalFitStatistics}}
}
\author{
Philipp Mueller
}
