% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_preprocessing.R
\name{generalExtremeExtractionInput}
\alias{generalExtremeExtractionInput}
\title{Extract the extreme events from a given time series.}
\usage{
generalExtremeExtractionInput()
}
\value{
menuItemOutput
}
\details{
Provides the shinydashboard::menuItemOutput for \code{\link{
generalExtremeExtraction}}. See the later one for details.
}
\author{
Philipp Mueller
}
\seealso{
Other preprocessing: \code{\link{cleaning.interactive}},
  \code{\link{deseasonalize.interactive}},
  \code{\link{deseasonalizeInput}},
  \code{\link{generalExtremeExtraction}}
}

