% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_fitting.R
\name{generalFittingRoutineInput}
\alias{generalFittingRoutineInput}
\title{Selection input to determine the fitting routine/method}
\usage{
generalFittingRoutineInput()
}
\value{
selectInput
}
\seealso{
Other climex-fitting: \code{\link{color.table}},
  \code{\link{data.fitting}}, \code{\link{data.initials}},
  \code{\link{fit.interactive}},
  \code{\link{generalFitStatisticsTable}},
  \code{\link{generalFitStatistics}}
}
\author{
Philipp Mueller
}
