% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{gev.density}
\alias{gev.density}
\title{Probability density function of GEV distribution}
\usage{
gev.density(parameters, z)
}
\arguments{
\item{parameters}{Fitted GEV parameters c( location, scale,
shape ).}

\item{z}{Numerical vector of sites where to evaluate the
density of the GEV distribution.}
}
\value{
Numerical vector of same length as z.
}
\description{
Calculate the probability density function for the
  generalized extreme value (GEV) distribution at a given point or
  at a number of points (provided as a numerical vector).
}
\details{
If you want to calculate the density of the Gumbel
  distribution, please stick to the c( location, scale, shape )
  scheme of the first argument while setting the shape parameter
  to zero. 

Port of the \code{\link[ismev]{gev.dens}} function.
}
\seealso{
Other extremes: \code{\link{block.list}},
  \code{\link{block.xts}}, \code{\link{block}},
  \code{\link{decluster.list}},
  \code{\link{decluster.xts}}, \code{\link{decluster}},
  \code{\link{extremal.index}}, \code{\link{gpd.density}},
  \code{\link{qevd}},
  \code{\link{return.level.climex.fit.gev}},
  \code{\link{return.level.climex.fit.gpd}},
  \code{\link{return.level.list}},
  \code{\link{return.level.numeric}},
  \code{\link{return.level}}, \code{\link{revd}},
  \code{\link{rlevd}}, \code{\link{threshold.list}},
  \code{\link{threshold.xts}}, \code{\link{threshold}},
  \code{\link{upper.limit.climex.fit.gev}},
  \code{\link{upper.limit.climex.fit.gpd}},
  \code{\link{upper.limit.list}},
  \code{\link{upper.limit.numeric}},
  \code{\link{upper.limit}}
}
\author{
Philipp Mueller
}
\concept{extremes}
