% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{gpd.density}
\alias{gpd.density}
\title{Probability density function of GP distribution}
\usage{
gpd.density(parameters, threshold, z)
}
\arguments{
\item{parameters}{Fitted GP parameters c( scale, shape )}

\item{threshold}{Threshold used to fit the GP distribution.}

\item{z}{Numerical vector of sites where to evaluate the
density of the GP distribution.}
}
\value{
Numerical vector of same length as z.
}
\description{
Calculate the probability density function for the
  generalized Pareto (GP) distribution at a given point or at a
  number of points (provided as a numerical vector).
}
\details{
If you want to calculate the density of the exponential
  distribution, please stick to the c( scale, shape ) scheme of
  the `parameters' argument while setting the shape parameter to
  zero. 

Port of the \code{\link[ismev]{gpd.dens}} function.
}
\seealso{
Other extremes: \code{\link{block.list}},
  \code{\link{block.xts}}, \code{\link{block}},
  \code{\link{decluster.list}},
  \code{\link{decluster.xts}}, \code{\link{decluster}},
  \code{\link{extremal.index}}, \code{\link{gev.density}},
  \code{\link{qevd}}, \code{\link{return.level}},
  \code{\link{revd}}, \code{\link{rlevd}},
  \code{\link{threshold.list}},
  \code{\link{threshold.xts}}, \code{\link{threshold}}
}
\author{
Philipp Mueller
}
