% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti.R
\name{likelihood.gradient}
\alias{likelihood.gradient}
\title{Calculates the gradient of the negative log likelihood of the GEV function.}
\usage{
likelihood.gradient(parameters, x.in)
}
\arguments{
\item{parameters}{Vector containing the location, scale and shape parameter.}

\item{x.in}{Time series.}
}
\value{
Numerical vector containing the derivative of the negative log likelihood in (location, scale, shape) direction.
}
\details{
Like \code{\{link{likelihood}} the quantity min.shape switches between the evaluation according to a Gumbel or a Frechet or Weibull like function. In the case of the Gumbel like type the gradient of the shape parameter is set to zero.
}
\author{
Philipp Mueller
}
\seealso{
Other optimization: \code{\link{fit.gev}},
  \code{\link{likelihood.initials}},
  \code{\link{likelihood}}
}

