% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti.R
\name{likelihood.initials}
\alias{likelihood.initials}
\title{Estimates the initial GEV or GPD parameters of a time series.}
\usage{
likelihood.initials(x, model = c("gev", "gpd"), type = c("best", "mom",
  "lmom"), modified = TRUE)
}
\arguments{
\item{x}{Time series/numeric.}

\item{model}{Determining whether to calculate the initial parameters of the GEV or GPD function. Default = "gev"}

\item{type}{Which method should be used to calculate the initial parameters. "best" combines all methods and return the result with the least likelihood. "mom" - method of moments returns an approximation according to the first two moments of the Gumbel distribution. "lmom" - returns an estimate according to the Lmoments method. Default = "best"}

\item{modified}{Determines if the skewness is getting used to determine the initial shape parameter. Default = TRUE.}
}
\value{
Numerical vector containing the c( location, scale, shape ) estimates for method = "gev" or the c( scale, shape ) estimates for method = "gpd".
}
\details{
Two main methods are used for the estimation: the L-moments method of Hosking & Wallis implemented in the extRemes::initializer.lmoments() function and an estimation using the first two moments of the Gumbel distribution. For the later one a modification was added: By looking at skewness of the distribution and with respect to some heuristic thresholds a shape parameter between -.4 and .2 is assigned. Warning: both methods do not work for samples with diverging (or pretty big) mean or variance. For this reason the restrict argument is included. If the estimates are bigger than the corresponding restrict.thresholds, they will be replaced by this specific value.
}
\author{
Philipp Mueller
}
\seealso{
Other optimization: \code{\link{fit.gev}},
  \code{\link{fit.gpd}}, \code{\link{likelihood.gradient}},
  \code{\link{likelihood}}
}

