% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti.R
\name{likelihood}
\alias{likelihood}
\title{Likelihood of the GEV and GP distribution}
\usage{
likelihood(parameters = NULL, x.in, model = c("gev", "gpd"))
}
\arguments{
\item{parameters}{Numerical vector containing the location, scale,
and shape parameters for the GEV or the scale and shape
parameters for the GPD. If NULL,
\code{\link{likelihood.initials}} is used to determine
them. Default = NULL}

\item{x.in}{Time series of class \pkg{xts}.}

\item{model}{String determining whether to calculate the initial
parameters of the GEV ("gev") or GPD ("gpd") function. Default =
"gev"}
}
\value{
Numerical value of the negative log likelihood.
}
\description{
Calculated the negative log likelihood of the GEV or
  GPD function.
}
\details{
This function is only meant to work with constant
  parameters and no covariates. \strong{x.in} is not called "x"
  since the call \code{grad( func = likelihood, x = parameters,
  ... )} wouldn't be possible.
}
\seealso{
Other optimization: \code{\link{fit.gev.list}},
  \code{\link{fit.gev.xts}}, \code{\link{fit.gev}},
  \code{\link{fit.gpd.list}}, \code{\link{fit.gpd.xts}},
  \code{\link{fit.gpd}},
  \code{\link{likelihood.augmented}},
  \code{\link{likelihood.gradient.augmented}},
  \code{\link{likelihood.gradient}},
  \code{\link{likelihood.initials}}
}
\author{
Philipp Mueller
}
