% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti.R
\name{likelihood}
\alias{likelihood}
\title{Calculated the negative log likelihood of the GEV function.}
\usage{
likelihood(parameters = NULL, x.in, verbose = FALSE)
}
\arguments{
\item{parameters}{Vector containing the location, scale and shape parameter. If NULL the \code{\link{likelihood.initials}} is used. Default = NULL}

\item{x.in}{Time series.}

\item{verbose}{Display debugging information.}
}
\value{
Numerical value of the negative log likelihood
}
\details{
This function is only meant to work with constant parameters and no covariats. x.in is not called "x" anymore since the call grad( func = likelihood, x = parameters, ... ) wouldn't be possible.
}
\author{
Philipp Mueller
}
\seealso{
Other optimization: \code{\link{fit.gev}},
  \code{\link{fit.gpd}}, \code{\link{likelihood.gradient}},
  \code{\link{likelihood.initials}}
}

