% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{multiplot}
\alias{multiplot}
\title{Plot multiple time series.}
\usage{
multiplot(plotlist = NULL, tt.title = main, main = NULL, ..., cols = 1,
  layout = NULL)
}
\arguments{
\item{plotlist}{Alternative way to pass \emph{ggplot} objects.}

\item{tt.title}{Same as \strong{main}.}

\item{main}{Title of the overall plot.}

\item{...}{Additional arguments for ggplot.}

\item{cols}{Number of columns in layout. Default = 1.}

\item{layout}{A matrix specifying the layout. If present,
\strong{cols} is ignored.}
}
\value{
print object.
}
\description{
This function takes several \pkg{ggplot2} plots and
  arranges them on one page.
}
\details{
Well, it is not really related to extreme value fitting
  and I am aware of the fact that its quite bad style to export
  auxiliary functions. But this one is just so handy. If the
  \strong{layout} is something like
  \code{matrix( c( 1, 2, 3, 3 ), nrow = 2, byrow = TRUE )},
  then plot 1 will go in the upper left, 2 will go in the upper
  right, and 3 will go all the way across the bottom. ggplot
  objects can be passed in ..., or to \strong{plotlist} (as a list
}
\examples{
anomalies.potsdam <- anomalies( temp.potsdam )
multiplot( anomalies.potsdam, temp.potsdam, cols = 2, main =
  "Difference between the pure data of the daily maximum
  temperatures of the Potsdam station and their anomalies" ) 
}
\seealso{
Other plot: \code{\link{ttplot}}
}
\author{
Paul Teetor (original), Philpp Mueller (minor changes)
}
