% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.climex.fit.gev}
\alias{plot.climex.fit.gev}
\title{Plots the GEV function fitted using \code{\link{fit.gev}}}
\usage{
\method{plot}{climex.fit.gev}(x)
}
\arguments{
\item{x}{Fitted GEV object.}
}
\value{
ggplot2 object.
}
\details{
Uses ggplot2
}

