% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.climex.fit.gev}
\alias{plot.climex.fit.gev}
\title{Plots the GEV function fitted using \code{\link{fit.gev}}}
\usage{
\method{plot}{climex.fit.gev}(x, bin.factor = NULL)
}
\arguments{
\item{x}{Fitted GEV object.}

\item{bin.factor}{Multiplying the length of x
by this factor, results in the number of bins
used in this plot. Default = NULL}
}
\value{
ggplot2 object.
}
\details{
Uses ggplot2. Since I will also use it
in the shiny app, where I want to adjust the number
of displayed bins, there is a second argument present.
}
