% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.climex.gev.fit}
\alias{plot.climex.gev.fit}
\title{Plots the GEV function fitted using \code{\link{gev.fit}}}
\usage{
\method{plot}{climex.gev.fit}(x)
}
\arguments{
\item{x}{Fitted GEV object.}
}
\value{
ggplot2 object.
}
\details{
Uses ggplot2
}

