% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{qevd}
\alias{qevd}
\title{Quantile calculation}
\usage{
qevd(p, location = NULL, scale = NULL, shape = NULL, threshold = NULL,
  model = c("gev", "gpd"), lower.tail = TRUE, silent = FALSE)
}
\arguments{
\item{p}{(Numeric) probability vector.}

\item{location}{Of the GEV distribution. Default = NULL.}

\item{scale}{Of the GEV/GP distribution. Default = NULL.}

\item{shape}{Of the GEV/GP distribution. Default = NULL.}

\item{threshold}{Used in the GP distribution. This parameter is
optional but should be provided in order to create a representation
of the fitted data exceedance. Default = NULL.}

\item{model}{Determines if to use the GEV or GP distribution.
Default = "gev".}

\item{lower.tail}{Whether to sample the probabilities P[X <= x] or
P[X > x]. Default = TRUE (first case).}

\item{silent}{Whether to display warnings or not. Default = FALSE.}
}
\value{
Numerical vector of the same length as input argument p.
}
\description{
Calculates the quantile of either the GEV or the GPD
distribution
}
\details{
Port from the extRemes package to (again) get rid of the
'threshold' argument to be able to have an separate 'threshold()'
function outside of the fitting function.
}
\seealso{
Other extremes: \code{\link{block.list}},
  \code{\link{block.xts}}, \code{\link{block}},
  \code{\link{decluster.list}},
  \code{\link{decluster.xts}}, \code{\link{decluster}},
  \code{\link{extremal.index}}, \code{\link{gev.density}},
  \code{\link{gpd.density}}, \code{\link{return.level}},
  \code{\link{revd}}, \code{\link{rlevd}},
  \code{\link{threshold.list}},
  \code{\link{threshold.xts}}, \code{\link{threshold}}
}
\author{
Philipp Mueller
}
