% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{remove.incomplete.years.list}
\alias{remove.incomplete.years.list}
\title{Remove incomplete years}
\usage{
\method{remove.incomplete.years}{list}(x, time.unit = 1, mc.cores = NULL)
}
\arguments{
\item{x}{A list of \pkg{xts} class objects.}

\item{time.unit}{Minimal differences found when applying the
\code{\link[base]{diff}} function to \strong{x}. Default = 1.}

\item{mc.cores}{A numerical input specifying the number of cores
to use for the multi core application of the function (see
\code{\link[parallel]{detectCores}}). This functionality is only
available if the input is a list of different objects. If NULL,
the function will be calculated classically with only one core.
Default = NULL.}
}
\value{
Object of the same class as the input \strong{x}.
}
\description{
Removes all years, which contain either a NA or are
  incomplete (derived from the amount of time stamps within a
  year).
}
\details{
Since incomplete years are detected via the
  difference in the time stamps of neighbouring points, the user
  has to provide the basic time unit. (Numerical value of the
  result of the \code{\link[base]{diff}} function applied to two
  consecutive points.) For daily data the value is 1.
}
\seealso{
\code{\link{remove.seasonality}}

Other ts: \code{\link{aic.climex.fit.gev}},
  \code{\link{aic.list}}, \code{\link{aic}},
  \code{\link{anomalies.list}},
  \code{\link{anomalies.xts}}, \code{\link{anomalies}},
  \code{\link{bic.climex.fit.gev}}, \code{\link{bic.list}},
  \code{\link{bic}}, \code{\link{check.completeness}},
  \code{\link{remove.incomplete.years.xts}},
  \code{\link{remove.incomplete.years}},
  \code{\link{remove.seasonality.list}},
  \code{\link{remove.seasonality.xts}},
  \code{\link{remove.seasonality}}
}
\author{
Philipp Mueller
}
