% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{remove.incomplete.years}
\alias{remove.incomplete.years}
\title{remove.incomplete.years}
\usage{
remove.incomplete.years(x)
}
\arguments{
\item{x}{Time series. Currently only works with the class xts since reference date in other objects is rather hard to get.}
}
\value{
Time series of class xts
}
\description{
Removes all years which contain either a NA or -999 or are incomplete. This procedure is done before extracting the seasonal component of the data.
}
\author{
Philipp Mueller
}
\seealso{
\code{\link{remove.seasonality}}

Other ts: \code{\link{aic}}, \code{\link{anomalies}},
  \code{\link{bic}}, \code{\link{remove.seasonality}},
  \code{\link{ts.swap}}
}

