% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{remove.incomplete.years}
\alias{remove.incomplete.years}
\title{Remove incomplete years}
\usage{
remove.incomplete.years(x, time.unit = 1)
}
\arguments{
\item{x}{Time series. Currently only works with the class xts
since reference date in other objects is rather hard to get.}

\item{time.unit}{Minimal differences found when applying the
`diff' function to `x'. Default = 1.}
}
\value{
Time series of class xts
}
\description{
Removes all years, which contain either a NA or are
  incomplete.
}
\details{
Since incomplete years are detected via the
  difference in the time stamps of neighbouring points, the user
  has to provide the basic time unit. (Numerical value of the
  result of the `diff' function applied to two consecutive
  points.) For daily data the value is 1.
}
\seealso{
\code{\link{remove.seasonality}}

Other ts: \code{\link{aic}}, \code{\link{anomalies}},
  \code{\link{bic}}, \code{\link{remove.seasonality}}
}
\author{
Philipp Mueller
}
