% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarCleaningInput}
\alias{sidebarCleaningInput}
\title{Toggling the cleaning of the time series.}
\usage{
sidebarCleaningInput()
}
\value{
menuItemOutput
}
\details{
Provides the shinydashboard::menuItemOutput for \code{\link{
sidebarCleaning}}. See the later one for details.
}
\author{
Philipp Mueller
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}}
}

