% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarDataBase}
\alias{sidebarDataBase}
\title{Selecting the database in the Climex app}
\usage{
sidebarDataBase(session)
}
\arguments{
\item{session}{Namespace session. For details check out
\url{http://shiny.rstudio.com/articles/modules.html}}
}
\value{
renderMenu
}
\description{
Selecting the database.
}
\details{
For now there are three different data sources you can
choose:the input provided when calling \code{\link{climex}}() on
  localhost or a file loaded via the sidebar, the station data of
  the German weather service (DWD) and artificial data sampled
  from a GEV distribution. When the climex app is used with the
  shiny-server, the input option will be disabled. Beware: it's
  not a true module! Using the namespaces does not really make
  sense because it has to be accessed from outside of this
  context.
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{file.loading}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarImprintInput}},
  \code{\link{sidebarImprint}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}},
  \code{\link{sidebarSeriesLengthInput}},
  \code{\link{sidebarSeriesLength}}
}
\author{
Philipp Mueller
}
