% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarDataBaseInput}
\alias{sidebarDataBaseInput}
\title{Selecting the database}
\usage{
sidebarDataBaseInput()
}
\value{
menuItemOutput
}
\details{
Provides the shinydashboard::menuItemOutput for \code{\link{
sidebarDataBase}}. See the later one for details.
}
\author{
Philipp Mueller
}
\seealso{
Other sidebar: \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}}
}

