% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climex_sidebar.R
\name{sidebarSeriesLengthInput}
\alias{sidebarSeriesLengthInput}
\title{Slider to determine the length of an artificial time series}
\usage{
sidebarSeriesLengthInput()
}
\value{
menuItemOutput
}
\details{
Provides the shinydashboard::menuItemOutput for \code{\link{
sidebarSeriesLength}}. See the later one for details
}
\seealso{
Other sidebar: \code{\link{data.selection}},
  \code{\link{file.loading}},
  \code{\link{sidebarCleaningInput}},
  \code{\link{sidebarCleaning}},
  \code{\link{sidebarDataBaseInput}},
  \code{\link{sidebarDataBase}},
  \code{\link{sidebarDataSourceInput}},
  \code{\link{sidebarDataSource}},
  \code{\link{sidebarDataTypeInput}},
  \code{\link{sidebarDataType}},
  \code{\link{sidebarLoadingGifOutput}},
  \code{\link{sidebarLoadingGif}},
  \code{\link{sidebarLoadingInput}},
  \code{\link{sidebarLoading}},
  \code{\link{sidebarSeriesLength}}
}
\author{
Philipp Mueller
}
