% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{source.data}
\alias{source.data}
\title{Load data of the DWD}
\usage{
source.data(pick.default = TRUE, download.path = NULL, envir = NULL)
}
\arguments{
\item{pick.default}{If TRUE it just search for the .RData file with
"default" in it's name and skips the interactive picking}

\item{download.path}{Specifies the data will be stored and
downloaded too. It is advised to store it in the path stored in
the options( "climex.path" ),  which is also used for importing
the saved data. You can overwrite its default value of
"~/R/climex/" by adding options( climex.path = "PATH" ) to your
.Rprofile path in your home.}

\item{envir}{Environment the data will be attached to. If
not specified, the data will be loaded to the environment the
function is called from. Default = NULL.}
}
\value{
Has no specific output but attaches .RData file to search path.
}
\description{
Access to the manually imported weather data.
}
\details{
Load .RData file generated by
\code{\link{download.data.dwd}}. This will be done interactively.
First all .RData files present in the download folder will be
printed. Than the user has to choose one of them by selecting its
position in the presented character vector via a numeric value
(e.g. 1 ). One can use the pick.default option to avoid the manual
selecting.
}
\author{
Philipp Mueller
}
