% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{source.data}
\alias{source.data}
\title{source.data}
\usage{
source.data(pick.default = TRUE, download.path = CLIMEX.PATH)
}
\arguments{
\item{pick.default}{If TRUE it just search for the .RData file with "default" in it's name and skips the interactive picking}

\item{download.path}{Specifies the data is stored.}
}
\value{
Has no specific output but attaches .RData file to search path.
}
\description{
Access to the manually imported weather data.
}
\details{
Load .RData file generated by \code{\link{download.data.dwd}}. This will be done interactively. First all .RData files present in the download folder will be printed. Than the user has to choose one of them by selecting its position in the presented character vector via a numeric value (e.g. 1 ). One can use the pick.default option to avoid the manual selecting.
}
\author{
Philipp Mueller
}

