% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{source.data}
\alias{source.data}
\title{Load a data file into R}
\usage{
source.data(download.path = NULL, envir = NULL)
}
\arguments{
\item{download.path}{Specifies the folder in which the function
will look for .RData files recursively. Per default the
\emph{R/climex/} directory in your home folder will be used. You
can overwrite 
this behavior by setting \code{options( climex.path = "PATH" )}
in your \emph{.Rprofile} path in your home.}

\item{envir}{Environment the data will be attached to. If
not specified, the data will be loaded to the environment the
function is called from. Default = NULL.}
}
\value{
Returns invisible but attaches the chosen .RData file to
  the specified R environment.
}
\description{
Searches the \emph{~/R/climex/} directory or a
  specified folder for .RData files recursively and displays the
  user its findings for her to choose one of them.
}
\details{
In order to use the data with the \pkg{climex}, the data
  should be of class \pkg{xts} or of lists of class \pkg{xts}
  objects.

  You can use the \pkg{dwd2r} package to download and use the
  daily station data provided by the German weather service.
}
\author{
Philipp Mueller
}
