% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ttplot}
\alias{ttplot}
\title{ttplot}
\usage{
ttplot(data.input, ...)
}
\arguments{
\item{data.input}{Time series which should be visualized.}

\item{...}{Additional parameters for the multiplot function.}
}
\value{
Nothing. The multiplot function is run in the last step.
}
\description{
Plotting a time series with ggplot2.
}
\details{
Plots all objects of class xts and ts in the same format but is also capable of handling lists of those objects and returning their plots using the \code{\link{multiplot}} function. With the main argument a title can be provided.
}
\author{
Philipp Mueller
}
\seealso{
Other plot: \code{\link{multiplot}}
}

