% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ttplot}
\alias{ttplot}
\title{Plotting a xts time series with ggplot2 in a convenient format.}
\usage{
ttplot(data.input, main = "Time series", ylab = NULL, x.df = NULL)
}
\arguments{
\item{data.input}{Time series which should be visualized.}

\item{...}{Additional parameters for the multiplot function.}
}
\value{
Nothing. The multiplot function is run in the last step.
}
\details{
Plots all objects of class xts. With the main argument a title can be provided.
}
\author{
Philipp Mueller
}
\seealso{
Other plot: \code{\link{multiplot}}
}

